### 2023-09-28 #5896 @editorialbot
[PRE REVIEW]: cblearn: Comparison-based Machine Learning in Python
**Submitting author:** @dekuenstle (<a href="http://orcid.org/0000-0001-5507-3731">David-Elias Künstle</a>)
**Repository:** https://github.com/cblearn/cblearn
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.1.2
