### 2023-10-04 706bfba
Add embedding dimension estimation routine (#64)
* Add dimension estimation with test
* Add matplotlib as test dependency
* Update changelog
