+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Upcoming
+
+- Feature: add routine for embedding dimension estimation to `embeddings` module (Künstle et al., 2022)
+- Fix: Avoid numpy deprecation warning for scalar variables in `fetch_similarity_matrix`
+
 ## 0.1.2
 
 - support python 3.11
