+++ setup.cfg
@@ -12,7 +12,7 @@ classifiers =
     Environment :: GPU :: NVIDIA CUDA
     Intended Audience :: Science/Research
     Intended Audience :: Developers
-    License :: OSI Approved :: MIT License  
+    License :: OSI Approved :: MIT License
     Programming Language :: Python
     Topic :: Software Development
     Topic :: Scientific/Engineering
@@ -56,11 +56,12 @@ tests =
     flake8~=6.0
     mypy~=1.1
     pandas~=1.5
+    matplotlib~=3.8
 docs =
     sphinx~=6.1
     sphinx_rtd_theme~=1.2
     sphinx-gallery~=0.12
-    matplotlib~=3.7
+    matplotlib~=3.8
 
 
 [build_sphinx]
