### 2023-10-05 f6eef64
Refactor embedding estimators to share fit in base
