+++ cblearn/embedding/tests/test_soe.py
@@ -2,9 +2,10 @@ import pytest
 import numpy as np
 from scipy.optimize import check_grad
 from scipy.stats import ttest_rel
-from cblearn.datasets import make_random_triplets
+from scipy.spatial import distance_matrix
 
-from cblearn.embedding._soe import SOE, _soe_loss
+from cblearn.datasets import make_random_triplets
+from cblearn.embedding import SOE
 
 
 def test_soe_repeat():
@@ -29,18 +30,28 @@ def test_soe_repeat():
     assert stat.pvalue < 0.05
 
 
-@pytest.mark.parametrize('n,d', [(20, 1), (50, 2), (100, 3)])
+@pytest.mark.parametrize('n,d', [(20, 1), (25, 2), (30, 3)])
 def test_soe_gradient(n, d):
     """ Test the gradient of the SOE loss function by comparing with local loss differences. """
-    def fun(x, *args):
-        return _soe_loss(x, *args)[0]
+    soe = SOE(n_components=d)
 
-    def grad(x, *args):
-        return _soe_loss(x, *args)[1]
+    def fun(x, x_shape, T, *args):
+        X = x.reshape(x_shape)
+        dist_pos = distance_matrix(X, X)[T[:, 0], T[:, 1]]
+        dist_neg = distance_matrix(X, X)[T[:, 2], T[:, 3]]
+        return soe._scipy_loss(X, T, dist_pos, dist_neg, *args)[0]
+
+    def grad(x, x_shape, T, *args):
+        X = x.reshape(x_shape)
+        dist_pos = distance_matrix(X, X)[T[:, 0], T[:, 1]]
+        dist_neg = distance_matrix(X, X)[T[:, 2], T[:, 3]]
+        return soe._scipy_loss(X, T, dist_pos, dist_neg, *args)[1]
 
     X = np.random.randn(n, d)
     T = make_random_triplets(X, size=2*d*n*np.log(n), result_format='list-order')
-    for i in range(10):  # test at 10 different points in the param space
-        init = 4 * np.random.rand(n, d) - 2  # [-2, 2]
-        args = [init.ravel(), X.shape, T[:, [1, 0, 0, 2]], 0.1]
-        assert check_grad(fun, grad, *args) < 1e-6
+
+    for ix_permut in ([1, 0, 0, 2], [0, 1, 0, 2], [0, 1, 2, 0], [1, 0, 2, 0]):
+        for i in range(5):  # test at different points
+            init = 4 * np.random.rand(n, d) - 2  # [-2, 2]
+            args = [T[:, ix_permut]]
+            assert check_grad(fun, grad, init.ravel(), init.shape, *args) < 1e-6
