### 2023-11-17 51273cc
Add image loading to car and things dataset
