+++ cblearn/datasets/_things_similarity.py
@@ -8,6 +8,8 @@ from os.path import join
 from typing import Optional, Union
 from urllib.request import urlretrieve
 import zipfile
+import scipy
+import warnings
 
 import numpy as np
 from sklearn.datasets import _base
@@ -39,6 +41,10 @@ def fetch_things_similarity(data_home: Optional[os.PathLike] = None, download_if
     ['aardvark', 'zucchini']
     >>> dataset.data.shape  # doctest: +REMOTE_DATA
     (146012, 3)
+    >>> dataset.embedding.shape  # doctest: +REMOTE_DATA
+    (1854, 49)
+    >>> dataset.images.shape  # doctest: +REMOTE_DATA
+    (1854, 150, 150, 3)
 
     Args:
         data_home : optional, default: None
@@ -61,6 +67,12 @@ def fetch_things_similarity(data_home: Optional[os.PathLike] = None, download_if
                 The first column is the selected odd-one.
             word : (n_objects,)
                 Single word associated with the thing objects.
+            uid : (n_objects,)
+                Unique word that identifies the images.
+            images: (n_objects, 150, 150, 3)
+                Preview images of the stimuli.
+            embedding : (n_objects, 49)
+                Sparse 49d embedding of the thing object based on the odd-one-out queries.
             synset : (n_objects,)
                 Wordnet Synset associated with the thing objects.
             wordnet_id : (n_objects,)
@@ -91,16 +103,31 @@ def fetch_things_similarity(data_home: Optional[os.PathLike] = None, download_if
         urlretrieve(ARCHIVE.url, archive_path)
 
         with zipfile.ZipFile(archive_path) as zf:
-            with zf.open('data/data1854_batch5_test10.txt', 'r') as f:
-                data = np.loadtxt(f, delimiter=' ')
+            with zf.open('variables/unique_id.txt', 'r') as f:
+                uid = np.loadtxt(f, delimiter=' ', dtype=str)
+
+            images, _ = _load_images(zf, uid)
+
+            with zf.open('data/spose_embedding_49d_sorted.txt', 'r') as f:
+                # explicitly state cols to avoid error due to trailing whitespace
+                embedding = np.loadtxt(f, delimiter=' ', usecols=range(0, 49))
+
+            data = _load_test_data(zf)
 
             with zf.open('items1854names.tsv', 'r') as f:
                 objects = np.array(list(csv.reader(io.TextIOWrapper(f), dialect='excel-tab'))[1:]).T
 
-        joblib.dump((data, objects), filepath, compress=6)
+        joblib.dump((data, objects, embedding, uid, images), filepath, compress=6)
         os.remove(archive_path)
     else:
-        (data, objects) = joblib.load(filepath)
+        cache = joblib.load(filepath)
+        if len(cache) == 5:
+            (data, objects, embedding, uid, images) = cache
+        else:  # backwards compatibility
+            (data, objects) = cache
+            embedding, uid, images = None, None, None
+            warnings.warn("Could not load embedding, uid and images from old cache file.\n"
+                          f"Consider cleaning cached file {filepath}.")
 
     if shuffle:
         random_state = check_random_state(random_state)
@@ -114,8 +141,44 @@ def fetch_things_similarity(data_home: Optional[os.PathLike] = None, download_if
         fdescr = rst_file.read()
 
     return Bunch(data=data,
+                 embedding=embedding,
+                 images=images,
+                 uid=uid,
                  word=objects[0],
                  synset=objects[1],
                  wordnet_id=objects[2],
                  thing_id=objects[5],
                  DESCR=fdescr)
+
+
+def _load_images(zf, uid):
[…]
