### 2023-11-17 a3d41de
Return triplet index in oddoneout preprocessing
