### 2023-11-17 208d079
Improve error messages in check_quadruplet
