+++ setup.cfg
@@ -12,7 +12,7 @@ classifiers =
     Environment :: GPU :: NVIDIA CUDA
     Intended Audience :: Science/Research
     Intended Audience :: Developers
-    License :: OSI Approved :: MIT License  
+    License :: OSI Approved :: MIT License
     Programming Language :: Python
     Topic :: Software Development
     Topic :: Scientific/Engineering
@@ -30,7 +30,7 @@ zip_safe = False
 include_package_data = True
 install_requires =
     numpy~=1.23
-    scipy~=1.10
+    scipy~=1.11
     scikit-learn~=1.2
     sparse~=0.14
 
