### 2023-12-04 #5896 @mbarzegary
Hi @ulf1 @sherbold @stsievert 👋
Considering that you have already published your software in JOSS, would any of you be willing to review this submission for JOSS? As you know, we carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Paper: https://raw.githubusercontent.com/openjournals/joss-papers/joss.05896/joss.05896/10.21105.joss.05896.pdf
Software: https://github.com/cblearn/cblearn
