### 2023-12-05 #5896 @sherbold
Hi @mbarzegary

While I am generally available as reviewer for JOSS, my schedule is brutal for the rest of this year. If a review in the first half of January is okay, then yes, otherwise I unfortunately need to decline.
