+++ cblearn/embedding/__init__.py
@@ -6,5 +6,5 @@ from cblearn.embedding._ste import STE, TSTE, MVTE
 from cblearn.embedding._oenn import OENN
 from cblearn.embedding._mlds import MLDS
 
-from cblearn.embedding._align import procrustes_alignment
+from cblearn.embedding._align import generalized_procrustes
 from cblearn.embedding._align import procrustes_standardize
