+++ paper/references.bib
@@ -56,6 +56,7 @@
 	author = {Roads, Brett D. and Love, Bradley C.},
 	year = {2021},
 	keywords = {Benchmark testing, Computer vision, Focusing, Measurement, Psychology, Semantics, Supervised learning},
+	doi = {10.1109/cvpr46437.2021.00355}
 }
 
 @article{jain_finite_2016,
@@ -291,7 +292,8 @@
   booktitle={Proceedings of the AAAI Conference on Human Computation and Crowdsourcing},
   volume={1},
   pages={69--77},
-  year={2013}
+  year={2013},
+  doi = {10.1609/hcomp.v1i1.13079}
 }
 
 @InProceedings{amid2015,
@@ -330,7 +332,8 @@
   booktitle={2018 IEEE International Conference on Acoustics, Speech and Signal Processing (ICASSP)},
   pages={3974--3978},
   year={2018},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1109/icassp.2018.8461868}
 }
 @inproceedings{ghosh2019landmark,
   title = {Landmark Ordinal Embedding},
