### 2024-01-08 cblearn/cblearn#66 @sherbold
Contributor installation instructions broken
The URL in the contributor installation instructions is broken and needs to be changed to: git@github.com:cblearn/cblearn.git

(cc https://github.com/openjournals/joss-reviews/issues/6139)
