### 2024-01-08 cblearn/cblearn#67 @sherbold
r_wrapper installation instructions missing
The ordinal embedding example cannot be executed without the r_wrapper. The installation for this is not documented and does not work out-of-the-box on a clean Ubuntu 22.04.

In general, there should be transparency for which features the r-wrapper is required, as well as further information what needs to be installed prior to calling `pip install cblearn[r_wrapper]`.

(cc https://github.com/openjournals/joss-reviews/issues/6139)
