### 2024-01-12 cblearn/cblearn#66 @haniyeka
I was going to mention the same as a minor issue. Thanks for bringing that up.
