### 2024-01-12 #6139 @haniyeka
I've had a close look at the cblearn package and how it fits into the Python ML ecosystem. Overall, the package is a valuable addition to the community, putting together comparison-based algorithms into a useful toolkit. I appreciate the efforts put into this and recommend its acceptance. But, I would like to highlight a few areas for improvement to maximize its potential:

- Partial Compatibility with Scikit-learn estimators:
  When I checked the test.yml workflow, it seems that 90 out of 301 tests are being skipped because cblearn's ordinal embedding estimators are not fully compatible with scikit-learn estimators. I'm not sure if this is something that can be addressed or not. The main issue here comes from cblearn’s approach to handle triplet comparisons as inputs, and this is different from the input data expected by scikit-learn’s estimators. This raises a bit of concerns about the generalizability of these estimators.

- Installation and Documentation:
  When I followed the installation instructions, I couldn't run the examples as the **h5py** package was missing. I would like to suggest the inclusion of an environment.yml/env.yml file in the cblearn package repository. This would improve the management of dependencies, make the setup process smoother.

- Certain Examples doesn't work:
  I wasnt able to run the ordinal_embedding.ipynb and triplet_formats.ipynb examples as mentioned in the issues. (https://github.com/cblearn/cblearn/issues/68 and https://github.com/cblearn/cblearn/issues/65)

- GPU Computation Feature Documentation:
  Although the package uses Torch for GPU computation, I couldn't find any information in the documentation about how users can enable this feature. This would be a nice addition to have.
