### 2024-01-15 79a65eb
Consider sample weight on soe optimization
