### 2024-01-15 270d460
Fix contributor install instructions  #66
