### 2024-01-16 02c7127
Do not use the wrapped estimator in the ordinal_embedding example #68
