### 2024-01-16 797135e
Fix triplet_format example  #65
