### 2024-01-16 30743a1
Run all examples when building docs  #65 #68
