### 2024-01-16 da04a4b
Fix warnings when building docs
