+++ docs/references/index.rst
@@ -79,7 +79,7 @@ Low-level Dataset Utility
 Utility
 -------
 
-.. automodule:: cblearn.embedding
+.. currentmodule:: cblearn.embedding
 
 .. autosummary::
     :toctree: generated/
@@ -143,7 +143,7 @@ Wrapper
     preprocessing.triplets_from_multiselect
     preprocessing.triplets_from_oddoneout
     preprocessing.triplets_from_mostcentral
-    preprocessing.SharedLabelEncoder
+    preprocessing.SharedColumnEncoder
     preprocessing.MultiColumnLabelEncoder
 
 :mod:`cblearn.utils` Utility
