### 2024-01-16 2b5cda3
Extend install instructions of extra requirements #67
