### 2024-01-16 d0b0e87
Fix docstring for dimension test
