### 2024-01-16 5eea0cf
Document when h5py is needed and raise better error if it is missing
