+++ cblearn/datasets/_imagenet_similarity.py
@@ -110,7 +110,14 @@ def fetch_imagenet_similarity(data_home: Optional[os.PathLike] = None, download_
         urlretrieve(ARCHIVE.url, archive_path)
 
         with zipfile.ZipFile(archive_path) as zf:
-            import h5py
+            try:
+                import h5py
+            except ImportError:
+                raise ImportError(
+                    "This function needs the extra package 'h5py' but could not find it. "
+                    "The package can be installed with pip install h5py."
+                    "On some platforms you might have to install hdf5 libraries separately.
+                    "")
 
             with zf.open('data/deprecated/psiz0.4.1/obs-118.hdf5', 'r') as f:
                 data_v1 = {k: np.asarray(v[()]) for k, v in h5py.File(f, mode='r').items()}
