### 2024-01-16 a52c7ce
Explain scikit learn compatibility in docs
