### 2024-01-16 #6139 @dekuenstle
@sherbold @haniyeka
 Thank you for your work on the detailed feedback in this thread and the issues.
  I am happy you wrote my submission as a valuable contribution.
   Your reviews have helped me to improve the documentation significantly.
Please find my point-by-point response to your issues below.

@stsievert, please take your time for the review, and don't feel pressured by this.
I'm getting ahead of myself and explaining the changes based on the previous reviews so that you don't have to deal with some issues again (e.g., the broken example). You can find the original submission in the `joss` branch. The changes here are in a separate `revision` branch, forked from the main.

* Scikit-learn compatibility.

   It is true that *cblearn* skips many of *sklearn*'s estimator tests.
    However, this is less due to the incompatibility of the estimator with the
    API than to the fact that *sklearn* generates artificial data in the tests incompatible with comparison data.
  I extended both the [User Guide](https://cblearn.readthedocs.io/en/revision/user_guide/index.html#scikit-learn-compatibility) and the [Contributor Guide](https://cblearn.readthedocs.io/en/revision/contributor_guide/index.html#scikit-learn-estimator-tests) accordingly:

    > All estimators in this library are compatible with the scikit-learn API and can be used in scikit-learn pipelines if comparisons are represented in the array format. The scikit-learn compatibility is achieved by implementing the fit, predict, and score methods of the BaseEstimator class. [...]

    > scikit-learn provides a test suite that should ensure the compatibility of estimators. We use this test suite to test our estimators, too, but have to skip some tests because they use artificial data incompatible to comparison data. Typically, cblearn estimators are compatible with scikit-learn estimators if comparisons are represented as numpy arrays. From an API perspective, comparison arrays look like discrete features and class labels; however, not all discrete features and class labels are valid comparisons.[...]

* Extra dependencies (h5py, r2py/wrapper, ...)

    cblearn/cblearn#67

    The most important dependencies of cblearn are installed by pip. However, more specialized functions require very large packages (*pytorch*) or, depending on the platform, require additional packages outside of Python (e.g., R interpreter).
    The extra dependencies required must be installed explicitly not to jeopardize platform compatibility.
     I do not want to bind the user to, for example, *conda* and therefore do not provide environment.yml. I know that we are entering into a tradeoff of user-friendliness here and have extended the [installation instructions](https://cblearn.readthedocs.io/en/revision/install.html#extra-requirements).

* Certain Examples don't work

    cblearn/cblearn#68
    cblearn/cblearn#65
    cblearn/cblearn#67

    I apologize for the inconvenience and have fixed the problems.
    To avoid such errors occurring again, the examples are now executed when building the documentation in the CD
    ([diff](https://github.com/cblearn/cblearn/pull/69/files#diff-85933aa74a2d66c3e4dcdf7a9ad8397f5a7971080d34ef1108296a7c6b69e7e3)).

* Contributor install broken

    cblearn/cblearn#66

    [I fixed the typo. ](https://cblearn.readthedocs.io/en/revision/install.html#contributor-installation)

* GPU Computation Feature Documentation

    I added a [section to the user guide](https://cblearn.readthedocs.io/en/revision/user_guide/index.html#pytorch-backend-cpu-gpu) that explains how and when to use the feature.

You can find the diff here:
https://github.com/cblearn/cblearn/pull/69
And the updated documentation is here:
https://cblearn.readthedocs.io/en/revision/
