### 2024-01-25 951d8f3
Improve documentation
* Update dependencies (+ fix broken test)
* Fix error in unit tests if  no remote data flag set
* Fix contributor install instructions  #66
* Do not use the wrapped estimator in the ordinal_embedding example #68
* Fix triplet_format example  #65
* Run all examples when building docs  #65 #68
* Fix warnings when building docs
* Extend install instructions of extra requirements #67
* Fix docstring for dimension test
* Document when h5py is needed and raise better error if it is missing
* Fix missing quote ending string literal
* Explain scikit learn compatibility in docs
