+++ docs/install.rst
@@ -19,38 +19,46 @@ The package is mainly tested on Linux, but Windows and Mac OS should work, too.
 User Installation
 -----------------
 
-``cblearn`` and its dependencies can be installed using `pip`:
+``cblearn`` and can be installed using `pip`:
 
 .. code-block:: bash
 
     pip install cblearn
 
+This will install the minimal set of required packages, sufficient for most uses and saving disk space.
+However, some features require more packages that can be installed by adding an ``option`` to the install command.
 
 .. _extras_install:
 
-Extras Requirements
+Extra Requirements
 ===================
 
-The installation routine above installs a minimal set of required packages, sufficient
-for most uses.
-However, some features require more packages that can be installed by adding 
-an `option` to the install command..
-
-For example, to use estimators on GPU, based on ``pytorch``, and estimators
-wrapping paper author's original implementation in ``R``-lang:
+Extra requirements can be installed by adding an ``option`` to the install command and enable more advanced features.
+Some of those extra dependencies need non-Python packages to be installed first.
 
 .. code-block:: bash
 
-    $ pip install cblearn[torch,wrapper]
+    $ pip install cblearn[torch,wrapper] h5py
 
-======= =============================================================
-Extras  Description
-======= =============================================================
-torch   Estimators which can run on a GPU, implemented using pytorch.
-wrapper Estimators which wrap implementations in R or Matlab/Octave.
-tests   Test runner for unit tests (required to contribute).
-docs    Build the documentation (required to contribute).
-======= =============================================================
+
+torch
+    Most estimators provide an (optional) implementation using ``pytorch`` to run large datasets on CPU and GPU.
+    This requires the ``pytorch`` package to be installed `manually <https://pytorch.org/get-started/locally/>`_
+    or by adding the ``torch``` option to the install command.
+    Note that ``pytorch`` might need about 1GB of disk space.
+
+wrapper
+    The estimators in :ref:`references_embedding_wrapper` provide an Python interface to the original implementation
+    in ``R``-lang.
+    This requires the ``rpy2`` package to be installed by adding the ``wrapper`` option to the install command.
+    Additionally, this requires an installed ``R`` interpreter whit must available be in the ``PATH`` environment variable.
+    The ``R`` packages are installed automatically upon the first use of the estimators.
+
+h5py
+    The function :func:`cblearn.datasets.fetch_imagenet_similarity` requires the ``h5py`` package to load the dataset.
+    This can package can be installed with pip.
+    Note that some platforms require additionally the ``hdf5`` libraries to be installed
+     `manually <https://www.hdfgroup.org/downloads/hdf5/>`_.
 
 
 .. _developer_install:
@@ -59,13 +67,33 @@ docs    Build the documentation (required to contribute).
 Contributor Installation
 ------------------------
 
+
 If you want to make changes to the code or documentation, you should
 first download the repository and install the project in developer mode with
-developer dependencies.
-This way, changes in the code are directly considered without the need for re-installation.
+developer dependencies.  This way, changes in the code are directly considered without the need for re-installation.
+Additionally, packages required to run the tests and build the documentation are installed.
 
 .. code-block:: bash
 
-    $ git clone git@github.com/cblearn/cblearn.git
+    $ git clone git@github.com:cblearn/cblearn.git
     $ cd cblearn
-    $ pip install -e.[tests,docs]
+    $ pip install -e.[tests,docs,torch,wrapper]
+
+The ``-e`` option installs the package in developer mode such that changes in the code are considered directly without re-installation.
+
+tests
+    To run the unit tests, the ``pytest`` package is required, which
+    can be installed by adding the ``tests`` option to the install command.
+
+docs
[…]
