+++ setup.cfg
@@ -31,11 +31,11 @@ packages = find:
 zip_safe = False
 include_package_data = True
 install_requires =
-    numpy~=1.24  # numba 0.57 requires numpy <1.25
+    numpy~=1.26  # numba 0.57 requires numpy <1.25
     scipy~=1.11
     scikit-learn~=1.3
-    sparse~=0.14
-    numba>=0.57  # force numba version supporting py3.11
+    sparse~=0.15
+    numba>=0.58  # force numba version supporting py3.11
 
 
 [options.package_data]
@@ -43,26 +43,25 @@ install_requires =
 
 [options.extras_require]
 torch =
-    torch~=2.0.0
+    torch~=2.1
 r_wrapper =
     rpy2~=3.5
 octave_wrapper =
     oct2py~=5.6
 tests =
-    pytest~=7.2
-    pytest-cov~=4.0
-    pytest-doctestplus~=0.12
+    pytest~=7.4
+    pytest-cov~=4.1
+    pytest-doctestplus~=1.1
     pytest-remotedata~=0.4
-    flake8~=6.0
-    mypy~=1.1
-    pandas~=1.5
+    flake8~=7.0
+    mypy~=1.8
+    pandas~=2.1
     matplotlib~=3.8
 docs =
-    sphinx~=6.1
-    sphinx_rtd_theme~=1.2
-    sphinx-gallery~=0.12
+    sphinx~=7.2
+    sphinx_rtd_theme~=2.0
+    sphinx-gallery~=0.15
     matplotlib~=3.8
-
 
 [build_sphinx]
 all-files = 1

[…]
