### 2024-01-25 780f6fe
Link stable documentation in README
