### 2024-01-25 cblearn/cblearn#66 @dekuenstle
Should be fixed now in the latest release: https://github.com/cblearn/cblearn/releases/tag/0.2.0
