### 2024-01-29 cblearn/cblearn#74 @stsievert
Docs are missing for some functions
The docs for [QueryScorer] and [query_error] are empty. What do those functions do?

[QueryScorer]:https://cblearn.readthedocs.io/en/revision/references/generated/cblearn.metrics.QueryScorer.html#cblearn.metrics.QueryScorer
[query_error]:https://cblearn.readthedocs.io/en/revision/references/generated/cblearn.metrics.query_error.html
