### 2024-01-29 cblearn/cblearn#75 @stsievert
PyTorch usage not clear
What's the reasoning behind "Note that pytorch might need about 1GB of disk space." on [the user guide][1].  When does that happen?

Is it possible to use PyTorch without GPUs? Is it possible to unset `CUDA_VISIBLE_DEVICES` and set `backend="pytorch"`?

[1]:https://cblearn.readthedocs.io/en/revision/user_guide/index.html#pytorch-backend-cpu-gpu
