### 2024-01-29 #6139 @stsievert
This is a decent software. This is software I might use, so I went through my review as if I were going to use it. I road-tested cblearn by creating an embedding with some of my personal data (input: object indices). Generally, this software provides good functionality – but I believe the documentation can be improved.

It took me a while to get the right input data format. I couldn't easily find adequate documentation. Eventually, I looked closely at the examples and essentially reproduced `preprocessing.query_from_columns`. An overview or "Getting Started" guide would have been really useful. Part of my problem was that I got confused by the `cblearn.utils` documentation. Why would a user read this documentation? It looks like an internal set of utility functions for the cblearn devs.

I have some other issues with the data input, echoing https://github.com/openjournals/joss-reviews/issues/6139#issuecomment-1889884329. The User Guide says a "response array containing 1 or -1 can specify if (i,j,k) is correct or wrong" but all the estimators in `cblearn.embedding` say `y` is ignored. I would encourage taking the human response variable `y` as input. Why can't `X` be a 2D array of indices and `y[i]` indicate if `X[i]` is in the expected format?

The paper makes some fairly bold claims about Scikit-learn compatibility. I am not sure if these are supported by the software. I might agree that relational data is different than feature data – but the [Scikit-learn estimator API][seapi] includes a "pairwise" tag, which might be relevant. I still have some concerns:

[seapi]:https://scikit-learn.org/stable/developers/develop.html#estimator-tags

- The input data is not verified:
    - It's possible for to pass floating point data to `predict` and get meaningless output (see below for script). If indices are required in `.fit`, shouldn't a ValueError be raised in `.predict` if an integer dtype is not received?
    - It's possible to pass any data into `.transform` and get the embedding back. Shouldn't there be some checks for array shape/size before returning the embedding? It's possible that `est.transform(X).shape == (100, 3)` when `X.shape == (5, 5)`, which doesn't make much sense.
- The checks in `test_sklearn_estimator_checks.py` have not been vetted; e.g., the nan check fails because embedding's don't "check for NaN and inf in transform" and the test for serialization/pickling is skipped even though it passes when the test isn't skipped.

<summary>Script for floating point predict input that produces meaningless result</summary>

``` python
from cblearn.embedding import SOE
import numpy as np

est = SOE(n_components=1)
n = 5  # n triplets
N = 100  # n responses
queries = [
    np.random.choice(n, size=3, replace=False)
    for _ in range(100)
]
responses = [
    [h, w, l] if abs(h - w) < abs(h - l) else [h, l, w]
    for h, w, l in queries
]

est.fit(responses)
assert est.embedding_.shape == (n, 1)

X = np.random.uniform(low=-1, high=1, size=(2 * N, 3))
y = est.predict(X)
assert y.shape == (2 * N, 3) and y.dtype.name == 'uint32'
assert np.allclose(y, 0)

```

There's also no statement of need in the documentation (though one does exist in the paper). Additionally, the paper is missing content:

- The paper references a section named "supplementary material" twice, but it's not available in the paper (it is available online, after knowing to navigate to Github). This material needs to be readily available (e.g., in the PDF or on the software documentation).
- There's some related work missing. Off the top of my head, [Salmon] (2) provides triplet "algorithms, datasets, automated tests and Scikit-learn integration" and [scikit-learn-contrib/metric-learn][metric-learn] (3) provides triplet embeddings with good Scikit-learn integration.

That covers most of the points in https://github.com/openjournals/joss-reviews/issues/6139#issuecomment-1892913087. I am very willing to accept this JOSS submission after these points are addressed.

-----

I also have some smaller questions/comments (in addition to https://github.com/cblearn/cblearn/issues/74 and https://github.com/cblearn/cblearn/issues/75):

- In the API section of your paper, it's claimed that cblearn "adapts [the scientific Python] conventions". Can you expand what you mean by "adapt"? That sentence would make a lot more sense with "adopt" -- is that a typo?
- In supplementary material: there's a typo, "our GPU (pytorch) implementations could not outperform the GPU ..." I think it should be "GPU ... CPU"
- "pytorch has a certain overhead (automatic derivation, data transformation)" on [the user guide]: that "certain overhead" is minimal I think, at least for the reasons you cite. PyTorch Tensors and NumPy arrays have very fast conversion (illustrated that in [a blog post of mine][pytorchspeed]), and the autodiff overhead is small (typically less than 1.5x) in this case (1). To that end, here's a script to use SciPy's minimize without a hard-coded gradient (e.g., similar to [cblearn/embeddings/soe.py][soe]) :

<details><summary>Script to use SciPy's optimize with PyTorch Tensors</summary>

``` python
import torch
import scipy.optimize
import numpy as np
from typing import Tuple

def f(x: np.ndarray, A: np.ndarray, y: np.ndarray) -> Tuple[float, np.ndarray]:
    A2 = torch.from_numpy(A)
    y2 = torch.from_numpy(y)

    x2 = torch.from_numpy(x)
    x2.requires_grad = True

    losses = (y2 - A2@x2)**2
    loss = losses.sum()
    loss.backward()

    return loss.item(), x2.grad.detach().numpy()

if __name__ == "__main__":
    n = 200  # n observations
    d = 100  # d features
    A = np.random.uniform(low=-1, high=1, size=(n, d))
    x_star = np.random.uniform(low=0, high=1, size=d)
    y = A @ x_star

    # Supported methods from scipy.optimize.minimize docstring:  ‘Newton-CG’, ‘trust-ncg’, ‘dogleg’, ‘trust-exact’, and ‘trust-krylov’
    result = scipy.optimize.minimize(
        f,
        x0=np.random.uniform(low=-0.1, high=0.1, size=d),
        args=(A, y),
        jac=True,
        method="Newton-CG",
    )

    norm = np.linalg.norm
    rel_error = norm(result.x - x_star) / norm(x_star)
    print(f"Relative error = {rel_error}")
```

References:

1. Griewank, Andreas. On automatic differentation. 1989.  https://ftp.mcs.anl.gov/pub/tech_reports/reports/P10.pdf
2. Sievert, Scott and Nowak, Rob and Rogers, Tim. 2022. https://joss.theoj.org/papers/10.21105/joss.04517
3. William de Vazelhes et al. metric-learn: Metric Learning Algorithms in Python. https://www.jmlr.org/papers/volume21/19-678/19-678.pdf

[soe]:https://github.com/cblearn/cblearn/blob/joss/cblearn/embedding/_soe.py
[metric-learn]:http://contrib.scikit-learn.org/metric-learn/
[Salmon]:https://docs.stsievert.com/stsievert/salmon
[pytorchspeed]:https://stsievert.com/blog/2017/09/07/pytorch/#numpyscipy-integration
[Scikit-learn tags]:https://scikit-learn.org/stable/developers/develop.html
[the user guide]:https://cblearn.readthedocs.io/en/revision/user_guide/index.html#pytorch-backend-cpu-gpu
[QueryScorer]:https://cblearn.readthedocs.io/en/revision/references/generated/cblearn.metrics.QueryScorer.html#cblearn.metrics.QueryScorer
[query_error]:https://cblearn.readthedocs.io/en/revision/references/generated/cblearn.metrics.query_error.html
