### 2024-02-12 7859b0f
Added synthetic data sampling with LinearSubspace as the first one
