+++ cblearn/datasets/_linear_subspace.py
@@ -1,0 +1,185 @@
+from cblearn.datasets._base import BaseManifold
+import numpy as np
+from scipy.stats import ortho_group
+from sklearn.utils import check_random_state
+from typing import Union, Dict, Callable
+from scipy.spatial.distance import pdist, squareform
+
+
+class LinearSubspace(BaseManifold):
+    """
+    Linear Subspace
+
+    Linear Subspace is a class for creating a hyperplane of a given
+    subspace dimension embedded in a higher dimensional space. It gives
+    a method of generating synthetic ordinal data with intrinsic
+    structure and dimensionality.
+
+    This class inherits from the BaseManifold class. This class creates
+    hyperplanes reproducibly using the scipy.stats.ortho_group function
+    for a given random state. This class can sample points from the
+    hyperplane using a given sampling function and add noise to the
+    points using a given noise function.
+
+    .. note:: Subspace dimension must be less than or equal to space
+    dimension and space dimension must be greater than 1.
+
+    Attributes:
+        subspace_dimension: Dimension of the subspace
+        space_dimension: Dimension of the space
+        random_state: Random state for reproducibility of the manifold
+        created: Flag to check if the hyperplane has been created
+        basis: Basis of the hyperplane
+
+    Examples:
+        >>> from cblearn.datasets import LinearSubspace, make_random_triplet_indices, triplet_response
+        >>> manifold = LinearSubspace(subspace_dimension=1, space_dimension=3)
+        >>> points, distances = manifold.sample_points(num_points=10)
+        >>> print(points.shape)
+        (10, 3)
+        >>> print(distances.shape)
+        (10, 10)
+        >>> # Generating points with noise
+        >>> noisy_points, noisy_distances = manifold.sample_points(10, noise='normal', noise_options={'scale': 0.1})
+        >>> # Responding to triplets based on distance matrix
+        >>> triplets = make_random_triplet_indices(n_objects=10, size=100)
+        >>> response = triplet_response(triplets, distances, distance='precomputed')
+    """
+    def __init__(self, subspace_dimension: int, space_dimension: int,
+                 random_state: Union[None, int, np.random.RandomState] = None):
+        """
+        Initialize the manifold
+
+        Args:
+            subspace_dimension: Dimension of the hyperplane
+            space_dimension: Dimension of the space in which the hyperplane
+                              is embedded
+            random_state: The seed of the pseudo random number generator
+                          to use when sampling. If None, the random number
+                          generator is the RandomState instance used by
+                          np.random.
+        """
+        if not isinstance(subspace_dimension, int):
+            raise ValueError('Subspace dimension must be an integer')
+        if subspace_dimension < 1:
+            raise ValueError('Subspace dimension cannot be less than 1')
+        if not isinstance(space_dimension, int):
+            raise ValueError('Space dimension must be an integer')
+        if subspace_dimension > space_dimension:
+            raise ValueError('Subspace dimension cannot be greater than'
+                             ' dimension')
+        if space_dimension <= 1:
+            raise ValueError('Space dimension cannot be less than 2')
+        self.subspace_dimension = subspace_dimension
+        self.space_dimension = space_dimension
+        random_state = check_random_state(random_state)
+        self.manifold_state = random_state
+        self.created = False
+        super().__init__(subspace_dimension=subspace_dimension,
+                         space_dimension=space_dimension,
+                         random_state=random_state)
+
+    def _create_manifold(self):
+        """ Creates the hyperplane """
+        # Source:
+        # https://stackoverflow.com/questions/69036765/sampling-random-points-from-linear-subspaces-of-a-given-radius-in-arbitary-dimen
+        if self.subspace_dimension == 1:
[…]
