+++ paper/references.bib
@@ -1,3 +1,27 @@
+@article{metric-learn,
+  title = {metric-learn: {M}etric {L}earning {A}lgorithms in {P}ython},
+  author = {{de Vazelhes}, William and {Carey}, CJ and {Tang}, Yuan and
+            {Vauquier}, Nathalie and {Bellet}, Aur{\'e}lien},
+  journal = {Journal of Machine Learning Research},
+  year = {2020},
+  volume = {21},
+  number = {138},
+  pages = {1--6}
+}
+
+@article{Sievert2023, doi = {10.21105/joss.04517}, url = {https://doi.org/10.21105/joss.04517}, year = {2023}, publisher = {The Open Journal}, volume = {8}, number = {84}, pages = {4517}, author = {Scott Sievert and Robert Nowak and Timothy Rogers}, title = {Efficiently Learning Relative Similarity Embeddings with Crowdsourcing}, journal = {Journal of Open Source Software} }
+
+@inproceedings{NIPS2015_89ae0fe2,
+ author = {Jamieson, Kevin G and Jain, Lalit and Fernandez, Chris and Glattard, Nicholas J. and Nowak, Rob},
+ booktitle = {Advances in Neural Information Processing Systems},
+ editor = {C. Cortes and N. Lawrence and D. Lee and M. Sugiyama and R. Garnett},
+ pages = {},
+ publisher = {Curran Associates, Inc.},
+ title = {NEXT: A System for Real-World Development, Evaluation, and Application of Active Learning},
+ url = {https://proceedings.neurips.cc/paper_files/paper/2015/file/89ae0fe22c47d374bc9350ef99e01685-Paper.pdf},
+ volume = {28},
+ year = {2015}
+}
 
 
 @article{vankadara_insights_2020,
