+++ paper/supplementary.md
@@ -36,9 +36,9 @@ The GPU implementations are slower on the tested datasets, presumably because th
 
 ## When should GPU implementations be preferred?
 
-In terms of accuracy and runtime, our GPU (or `pytorch`) implementations could not outperform the GPU (or `scipy`) pendants on the tested datasets. However, \autoref{fig:performance-per-algorithm_cblearn} shows the GPU runtime grows slower with the number of triplets, such that they potentially outperform CPU implementations with large datasets of $10^7$ triplets and more. In some cases, the GPU implementations show the overall best accuracy.
-An additional advantage of GPU implementations is that they require no explicit gradient definition, which allows easier tweaking of algorithms.
-However, the stochastic optimization process of these GPU implementations might be more sensitive to hyperparameter choices.
+In terms of accuracy and runtime, our GPU (or `pytorch`) implementations could not outperform the CPU (or `scipy`) pendants on the tested datasets. However, \autoref{fig:performance-per-algorithm_cblearn} shows the GPU runtime grows slower with the number of triplets, such that they potentially outperform CPU implementations with large datasets of $10^7$ triplets and more. In some cases, the GPU implementations show the overall best accuracy.
+An additional advantage of GPU implementations is that they require no explicit gradient definition, which simplifies the implementation of new algorithms.
+However, the stochastic optimization process of these GPU implementations might be more sensitive to hyperparameter choices and thus requires more tuning. 
 
 ![The runtime increases almost linearly with the number of triplets. However, GPU implementations have a flatter slope and thus can compensate for the initial time overhead on large datasets.
     \label{fig:time-per-triplets_gpu}](images/time-per-triplets_gpu.pdf){width=50%}
