### 2024-02-14 d441517
Improve docstrings, add missing docstrings #74 #42
