### 2024-02-14 006197a
Add validation for triplet data in predictions
