### 2024-02-14 0cf2f49
Raise value error if queries are passed to .transform
