+++ cblearn/metrics/_triplets.py
@@ -1,12 +1,6 @@
 import numpy as np
 from sklearn.utils import check_array
 from sklearn import metrics
-try:
-    from sklearn.metrics._scorer import _Scorer
-    _legacy_scorer = False
-except ImportError:
-    from sklearn.metrics._scorer import _PredictScorer as _Scorer
-    _legacy_scorer = True
 import sparse
 import scipy
 
@@ -70,17 +64,11 @@ def query_error(true_response: utils.Response, pred_response: utils.Response) ->
     return 1 - query_accuracy(true_response, pred_response)
 
 
-class QueryScorer(_Scorer):
-    """Scorer class for query accuracy.
+def query_accuracy_scorer(clf, X, y):
+    """Scorer function for query accuracy, compatible with sklearn's scorer API.
 
     See :py:func:`cblearn.metrics.query_accuracy` for more information.
     """
-    def __init__(self):
-        if _legacy_scorer:
-            super().__init__(QueryScorer._score_func, 1, {})
-        else:
-            super().__init__(QueryScorer._score_func, 1, {}, 'predict')
-
-    def _score_func(true_response, query):
-        query, pred_response = utils.check_query_response(query, result_format='list-boolean')
-        return query_accuracy(true_response, pred_response)
+    X, y = utils.check_query_response(X, y, result_format='list-count')
+    y_pred = clf.predict(X, result_format='list-count')
+    return query_accuracy(y, y_pred)
