+++ cblearn/embedding/_base.py
@@ -9,28 +9,25 @@ from cblearn import utils
 from cblearn import metrics
 
 
-class TripletEmbeddingMixin(TransformerMixin):
+class TripletEmbeddingMixin():
     def _more_tags(self):
         return {
             'requires_positive_X': True,
             'requires_positive_y': True,
             'X_types': ['categorical'],
+            'binary_only': True,  # enforce binary y in tests
+            'triplets': True  # enforce triplet X in tests
         }
 
     def fit_transform(self, X: utils.Query, y: Optional[np.ndarray] = None, **fit_params) -> np.ndarray:
         if y is None:
             # fit method of arity 1 (unsupervised transformation)
-            return self.fit(X, **fit_params).transform()
+            self.fit(X, **fit_params)
+            return self.embedding_
         else:
             # fit method of arity 2 (supervised transformation)
-            return self.fit(X, y, **fit_params).transform()
-
-    def transform(self, X: Optional[utils.Query] = None, y: Optional[np.ndarray] = None) -> np.ndarray:
-        """ Use .fit_transform instead. """
-        check_is_fitted(self, 'embedding_')
-        if X is not None:
-            raise ValueError(f"Expected no input data, got {X}.")
-        return self.embedding_
+            self.fit(X, y, **fit_params)
+            return self.embedding_
 
     def predict(self, X: utils.Query, result_format: Optional[utils.Format] = None) -> np.ndarray:
         check_is_fitted(self, 'embedding_')
@@ -48,6 +45,5 @@ class TripletEmbeddingMixin(TransformerMixin):
         Returns.
             Fraction of correct triplets.
         """
-        if y is None:
-            y = X
-        return metrics.query_accuracy(self.predict(X), y)
+        X, y = utils.check_query_response(X, y, result_format='list-count')
+        return metrics.query_accuracy(self.predict(X, result_format='list-count'), y)
