+++ cblearn/embedding/_soe.py
@@ -122,6 +122,7 @@ class SOE(BaseEstimator, TripletEmbeddingMixin):
             self.
         """
         queries = utils.check_query_response(X, y, result_format='list-order')
+        self.n_features_in_ = 3
         if not n_objects:
             n_objects = queries.max() + 1
         random_state = check_random_state(self.random_state)
