### 2024-02-15 4db2198
Extend user guide about triplets and pytorch
