+++ docs/user_guide/index.rst
@@ -6,8 +6,11 @@ User Guide
 
 Most Machine Learning algorithms use numerical training data (features) for inference,
 either representing points in a Euclidean space, similarities, or distances.
-The are settings, e.g. in human studies, when metric points are not available but only ordinal comparisons.
-Comparison-based learning algorithms are the machine learning algorithms applicable in this setting.
+There are settings, e.g., in human studies, when obtaining featurized data is difficult, but comparisons are straightforward.
+Comparison-based learning algorithms are the machine learning algorithms applicable in this setting,
+as they learn from ordinal comparisons between object similarity.
+
+.. _triplet_comparison:
 
 -------------------
 Triplet comparisons
@@ -22,10 +25,64 @@ inequality:
 
     \delta(x_i, x_j) \le \delta(x_i, x_k).
 
-This library supports two representation formats of triplets in an array or a sparse matrix form.
-The array form uses 2d ``numpy`` arrays representing a triplet per row and columns for ``i,j,k``.
-Alternatively to the ordering, an additional response array containing 1 or -1 can specify if  ``(i,j,k)`` is correct or wrong.
-The sparse matrix is an alternative representation, where triplets are naturally specified as the matrix indices, containing entries 1 or -1.
+Triplets can represented as 2d ``numpy`` arrays, where each row represents a triplet and the columns represent the indices of the objects.
+In the library, we call this representation the ordered list format or ``'list-order'``.
+
+.. code-block:: python
+
+    import numpy as np
+    triplet_responses = np.array([[0, 1, 2],  # 0 is closer to 1 than 2
+                                  [2, 3, 1],
+                                  [1, 3, 0]])
+
+
+In some applications, having a separate array containing the correct response for each triplet is helpful.
+This array contains 1 for correct triplets and -1 for incorrect triplets.
+The following arrays in the ``'list-count'`` format are equivalent to the previous example.
+
+.. code-block:: python
+
+    triplets = np.array([[0, 1, 2],
+                         [2, 1, 3],
+                         [1, 0, 3]])
+    responses = np.array([1, -1, -1])
+
+These array-based formats are handy in interacting with other machine learning procedures as every row represents a "data point".
+Therefore, it is straightforward to sample triplets or split the data into training and test sets.
+
+.. code-block:: python
+
+    from sklearn.model_selection import train_test_split
+    X_train, X_test, y_train, y_test = train_test_split(triplets, responses, test_size=0.2)
+
+However, if you want to calculate directly with triplet data, the ``'sparse'`` representation is more suitable.
+Here, we use a sparse array with three dimensions and as many entries as there are objects.
+Based on the position in the array, entries with -1 and 1 encode whether the triplet is correct or incorrect.
+
+.. code-block:: python
+
+    import sparse
+    sp_triplets = sparse.COO(triplets.T, responses, shape=(4, 4, 4))
+    sp_triplets[0, 1, 2]  # returns  1
+    sp_triplets[2, 1, 3]  # returns -1
+    sp_triplets[1, 0, 3]  # returns -1
+    sp_triplets[0, 1, 3]  # returns  0
+
+    new_sp_triplets = sp_triplets + another_sp_triplets  # element-wise addition to join two datasets
+
+
+The triplet-based algorithms in the ``cblearn`` library typically can handle all of the previous.
+An ordinal embedding estimator, for example, can train with either of the following:
+
+.. code-block:: python
+
+    from cblearn.embedding import SOE
+
+    soe = SOE(n_components=2)
+    soe.fit(triplet_responses)
+    soe.fit(triplets, responses)
+    soe.fit(sp_triplets)
+
 
 --------------------------
 Scikit-learn compatibility
[…]
