### 2024-02-15 ef91f90
Extend supplementary with pytorch toy example
