+++ paper/supplementary.md
@@ -38,10 +38,23 @@ The GPU implementations are slower on the tested datasets, presumably because th
 
 In terms of accuracy and runtime, our GPU (or `pytorch`) implementations could not outperform the CPU (or `scipy`) pendants on the tested datasets. However, \autoref{fig:performance-per-algorithm_cblearn} shows the GPU runtime grows slower with the number of triplets, such that they potentially outperform CPU implementations with large datasets of $10^7$ triplets and more. In some cases, the GPU implementations show the overall best accuracy.
 An additional advantage of GPU implementations is that they require no explicit gradient definition, which simplifies the implementation of new algorithms.
-However, the stochastic optimization process of these GPU implementations might be more sensitive to hyperparameter choices and thus requires more tuning. 
 
 ![The runtime increases almost linearly with the number of triplets. However, GPU implementations have a flatter slope and thus can compensate for the initial time overhead on large datasets.
     \label{fig:time-per-triplets_gpu}](images/time-per-triplets_gpu.pdf){width=50%}
+
+There are various explanations for the speed disadvantage of our pytorch implementations. On the one hand, it may be due to the overhead of converting between numpy and pytorch and calculating the gradient (AutoGrad). On the other hand, it can also be due to the optimizer or the selected hyperparameters. 
+To get a first impression of these factors, we have built a toy example, linear regression with 200 observations and 100 dimensions. The \autoref{fig:torch-speedtest} shows that the overhead of autograd and the stochastic optimization (Adam, lr=0.05) both slow down the optimization multiplicatively by factor ~8 in this example. However, it can be assumed, and in accordance with the above tendency, that this disadvantage decreases with increasing data set size. 
+
+![The runtime and error for different optimization methods in a toy example.\label{fig:torch-speedtest}](
+    images/torch_speedtest.pdf
+){width=50%}
+
+An additional disadvantage of stochastic optimizers like Adam is, that they are more sensitive to hyperparameter choices and thus require more tuning. This sensitivity is demonstrated in \autoref{fig:adam_lr}, where the learning rate of Adam is varied for the toy example. Especially runtime largely depends on the learning rate, while the error is less sensitive to it. Likewise, the performance of `pytorch` ordinal embedding implementations could be improved by using more sophisticated tuning of optimizer parameters.
+
+
+![The runtime and error for different learning rates of the Adam optimizer in a toy example.\label{fig:adam_lr}](
+    images/adam_lr.pdf
+){width=50%}
 
 ## How does `cblearn` compare to other implementations?
 
