### 2024-02-15 42a883b
Add more references that used the library, acknowledge reviewers and editor
