+++ paper/references.bib
@@ -9,6 +9,23 @@
   pages = {1--6}
 }
 
+@article{sauer2023seeing,
+  title={Seeing the future of progressive glasses: a new perceptual approach to spectacle lens design},
+  author={Sauer, Yannick and K{\"u}nstle, David-Elias and Wichmann, Felix and Wahl, Siegfried},
+  year={2023},
+  publisher={PsyArXiv},
+  doi={10.31234/osf.io/pge5n}
+}
+
+@article{huber2024tracing,
+  title={Tracing Truth Through Conceptual Scaling: Mapping People’s Understanding of Abstract Concepts},
+  author={Huber, Lukas S and K{\"u}nstle, David-Elias and Reuter, Kevin},
+  year={2024},
+  publisher={PsyArXiv},
+  doi={10.31234/osf.io/c42yr}
+}
+
+
 @article{Sievert2023, doi = {10.21105/joss.04517}, url = {https://doi.org/10.21105/joss.04517}, year = {2023}, publisher = {The Open Journal}, volume = {8}, number = {84}, pages = {4517}, author = {Scott Sievert and Robert Nowak and Timothy Rogers}, title = {Efficiently Learning Relative Similarity Embeddings with Crowdsourcing}, journal = {Journal of Open Source Software} }
 
 @inproceedings{NIPS2015_89ae0fe2,
