### 2024-02-15 9864f6b
Minor fix to run tests on 3.11
