### 2024-02-15 #6139 @dekuenstle
Dear @stsievert,
thank you for your review and the great recommendations. I think they helped me to considerably improve the unit tests, the documentation for new users and the paper. I'm glad that you like the functionality and agree that the documentation is a notable weakness–so I have significantly revised it. You can find the details, point by point, below. I hope I have addressed all your issues, and I welcome your feedback.

> It took me a while to get the right input data format. I couldn't easily find adequate documentation. Eventually, I looked closely at the examples and essentially reproduced `preprocessing.query_from_columns`. An overview or "Getting Started" guide would have been really useful.
> Part of my problem was that I got confused by the `cblearn.utils` documentation. Why would a user read this documentation? It looks like an internal set of utility functions for the cblearn devs.
>
> I have some other issues with the data input, echoing [#6139 (comment)](https://github.com/openjournals/joss-reviews/ issues/6139#issuecomment-1889884329). The User Guide says a "response array containing 1 or -1 can specify if (i,j,k) is correct or wrong" but all the estimators in `cblearn.embedding` say `y` is ignored. I would encourage taking the human response variable `y` as input. Why can't `X` be a 2D array of indices and `y[i]` indicate if `X[i]` is in the expected format?

I agree that it was too hard to get started as a user due to the documentation.
Therefore, I added  *welcome* and *getting started* sections and significantly extended the *user guide*.
In addition, I extended the documentation of some functions, for example, of `preprocessing.query_from_columns`.
Please find the new documentation here: https://cblearn.readthedocs.io/en/joss

The *getting started*, and the user guide should now introduce the data formats more appropriately. Indeed it is possible but optional to pass response variables to `y` as input. I, however, have to admit that the great flexibility of input formats in *cblearn* can indeed be confusing, mainly due to implicit conversions. Therefore I am planning some changes in the data format in the future, which is more restrictive and requires more explicit conversations. If you are interested, you can find work in progress in this [branch](https://github.com/cblearn/cblearn/blob/79a65eb410845a953265b21e849821f1d98fc4e6/cblearn/embedding/_base.py#L81C1-L81C65).

> The paper makes some fairly bold claims about Scikit-learn compatibility. I am not sure if these are supported by the software.
> I might agree that relational data is different than feature data – but the [Scikit-learn estimator API](https://scikit-learn.org/stable/developers/develop.html#estimator-tags) includes a "pairwise" tag, which might be relevant.
> The checks in `test_sklearn_estimator_checks.py` have not been vetted; e.g., the nan check fails because embedding's don't "check for NaN and inf in transform" and the test for serialization/pickling is skipped even though it passes when the test isn't skipped.

I agree, that estimator tests should ensure compatibility with sklearn. Therefore, I worked on them again such that all of them run through (0 skipped).
Thank you for the hint with the "pairwise" tag: While I couldn't use this tag directly (it's intended for estimators that require squared input matrixes), I introduced a "triplets" tag. Based on this tag, I now monkey-patch the data generation of sklearn's estimator tests.

In addition to these changes in the code, I changed the respective sentence in the paper to clarify that compatibility does not hold for all data formats but just the numpy array format.

>     * The input data is not verified:
>
>       * It's possible for to pass floating point data to `predict` and get meaningless output (see below for script). If indices are required in `.fit`, shouldn't a ValueError be raised in `.predict` if an integer dtype is not received?
>       * It's possible to pass any data into `.transform` and get the embedding back. Shouldn't there be some checks for array shape/size before returning the embedding? It's possible that `est.transform(X).shape == (100, 3)` when `X.shape == (5, 5)`, which doesn't make much sense.

Another reason of the planned restrictions of input types (see above), is to do better data verification in the future.
These two cases, however, can already be verified today. I removed the `.transform` method from embedding estimators and verify the triplet input in a method used by  `.predict` (see the [unit test](https://github.com/cblearn/cblearn/blob/9c7af5d5791881a6eae717581fc939938a94f7d0/cblearn/datasets/tests/test_triplet_response.py#L7)).

> There's also no statement of need in the documentation (though one does exist in the paper).

I've included the statement of need as part of the new welcome page.

>     * The paper references a section named "supplementary material" twice, but it's not available in the paper (it is available online, after knowing to navigate to Github). This material needs to be readily available (e.g., in the PDF or on the software documentation).

In the paper, I've linked the supplementary material to the PDF in the main branch (this link does not work yet, but will work after the merge of the *joss* branch to *main*):
[Supplementary Material](https://github.com/cblearn/cblearn/blob/main/paper/supplementary.pdf)

>     * There's some related work missing. Off the top of my head, [Salmon](https://docs.stsievert.com/stsievert/salmon) (2) provides triplet "algorithms, datasets, automated tests and Scikit-learn integration" and [scikit-learn-contrib/metric-learn](http://contrib.scikit-learn.org/metric-learn/) (3) provides triplet embeddings with good Scikit-learn integration.

Thank you for pointing out these other works. I extended the related work section to explicitely mention metric-learn, NEXT & salmon. Additionally, I added a related work section to the README.

        >> Related packages with collections of comparison-based learning algorithms  have a focus on metric learning [`metric-learn` with a high compatibility to scikit-learn, @metric-learn] and crowd-sourced data collection, using active ordinal embedding algorithms [e.g., `NEXT`, @NIPS2015_89ae0fe2 and `salmon`, @Sievert2023]. Our package `cblearn`, on the other hand, focuses on providing comparison data and interoperable estimator implementations of the remaining areas of comparison-based learning.

>     * In the API section of your paper, it's claimed that cblearn "adapts [the scientific Python] conventions". Can you expand what you mean by "adapt"? That sentence would make a lot more sense with "adopt" -- is that a typo?
>     * In supplementary material: there's a typo, "our GPU (pytorch) implementations could not outperform the GPU ..." I think it should be "GPU ... CPU"

Thank you. Both typos are fixed now.

>     * "pytorch has a certain overhead (automatic derivation, data transformation)" on [the user guide](https://cblearn.readthedocs.io/en/revision/user_guide/index.html#pytorch-backend-cpu-gpu): that "certain overhead" is minimal I think, at least for the reasons you cite. PyTorch Tensors and NumPy arrays have very fast conversion (illustrated that in [a blog post of mine](https://stsievert.com/blog/2017/09/07/pytorch/#numpyscipy-integration)), and the autodiff overhead is small (typically less than 1.5x) in this case (1). To that end, here's a script to use SciPy's minimize without a hard-coded gradient (e.g., similar to [cblearn/embeddings/soe.py](https://github.com/cblearn/cblearn/blob/joss/cblearn/embedding/_soe.py)) :

I agree, that the documentation was not very specific in this case and that the conversation is not the main problem.
Based on your example, I run an a comparison between scipy, scipy+AutoGrad, and torch optimization on a toy problem and found, that the main "problem" is the Adam optimization ( [code](https://github.com/cblearn/cblearn-benchmark/blob/main/scripts/pytorch_speedtest.ipynb)). I extended the paper's supplementary and the documentation's user guide accordingly.

![Optimization speed test on toy example.](https://raw.githubusercontent.com/cblearn/cblearn/joss/docs/user_guide/torch_speedtest.png)
