### 2024-02-20 a8400ce
Reinsert tripletembedding.transform, ensure that checks are passing.
