+++ cblearn/embedding/_forte.py
@@ -87,6 +87,7 @@ class FORTE(BaseEstimator, TripletEmbeddingMixin):
         Returns:
             self.
         """
+        self.fit_X_ = utils.check_query(X, result_format='list-order')  # for data validation in .transform
         triplets = utils.check_query_response(X, y, result_format='list-order')
         self.n_features_in_ = 3
         if not n_objects:
