+++ cblearn/embedding/_ste.py
@@ -110,6 +110,7 @@ class STE(BaseEstimator, TripletEmbeddingMixin):
         Returns:
             self.
         """
+        self.fit_X_ = utils.check_query(X, result_format='list-order')  # for data validation in .transform
         triplets = utils.check_query_response(X, y, result_format='list-order')
         self.n_features_in_ = 3
         if not n_objects:
