+++ cblearn/datasets/tests/test_triplet_response.py
@@ -13,10 +13,14 @@ def test_triplet_response_validates_input():
         for _ in range(t)
     ]
     invalid_queries_1 = [
+        np.random.choice(n, size=5, replace=False)
+        for _ in range(t)
+    ]
+    invalid_queries_2 = [
         np.random.choice(n + 1, size=3, replace=False)
         for _ in range(t)
     ]
-    invalid_queries_2 = np.random.uniform(low=-1, high=1, size=(t, 3))
+    invalid_queries_3 = np.random.uniform(low=-1, high=1, size=(t, 3))
     embedding = np.random.normal(size=(n, d))
 
     responses = triplet_response(valid_queries, embedding)
@@ -25,3 +29,5 @@ def test_triplet_response_validates_input():
         triplet_response(invalid_queries_1, embedding)
     with pytest.raises(ValueError):
         triplet_response(invalid_queries_2, embedding)
+    with pytest.raises(ValueError):
+        triplet_response(invalid_queries_3, embedding)
