### 2024-02-28 6285963
Use a more memory-efficient method to calculate the number of unique values in triplets
