### 2024-02-29 996a763
Update plots in supplementary to show the individual data points
