### 2024-03-07 #6139 @stsievert
Thanks for the updates @dekuenstle. I've checked off most of the boxes in https://github.com/openjournals/joss-reviews/issues/6139#issuecomment-1892913087, but I have one last suggestion on the graphs in the supplementary material:

> I run an a comparison between scipy, scipy+AutoGrad, and torch optimization on a toy problem

Thanks for that comparison and the graph. That graph looks really interesting, at least at first. I think that toy problem is inadequate, it's just noiseless least squares with not much data. I think that graph would be much better if it actually used triplet losses and datasets. Obtaining the "SciPy + Autograd" comparison is pretty straightforward for triplets:

<summary>Example script to minimizing CKL loss</summary>

This script defines a decorator `pytorch_grad`. That allows for pretty straightforward use of SciPy's `minimize`. after moving the optimization function NumPy to Torch (see below for example). Then this script spends time defining a dataset before calling `minimize`.

The `_ckl_x_loss` function below is basically copy/pasted from [_ckl.py#L143-L153][ckl], with some small modifications.

[ckl]:https://github.com/cblearn/cblearn/blob/eecf85f9fde0bf264041a323623b630dbdfd5aa2/cblearn/embedding/_ckl.py#L143-L153

``` python
import itertools
import torch
import numpy as np
from scipy.optimize import minimize

def pytorch_grad(f):
    def get_grad(x, *args):
        y = torch.from_numpy(x)
        y.requires_grad = True
        args = list(args)
        for i, a in enumerate(args):
            if isinstance(a, np.ndarray):
                args[i] = torch.from_numpy(a)
        loss = f(y, *args)
        loss.backward()
        grad = y.grad
        return loss, grad  #.ravel()
    return get_grad

@pytorch_grad
def _ckl_x_loss(x, x_shape, triplets, mu, float_min=np.finfo(float).tiny):
    X = x.reshape(x_shape)

    # Perform the expansion of (x - y)**2 = x@x + y@y - 2x@y w/ broadcasting
    # (more easily allows the gradient to be tracked)
    _Dxy = X @ X.T
    _Dx = (X**2).sum(axis=1).reshape(1, -1)
    _Dy = (X**2).sum(axis=1).reshape(-1, 1)
    D = _Dx + _Dy - 2*_Dxy

    I, J, K = tuple(triplets.T)
    nom = mu + D[I, K]
    den = 2 * mu + D[I, K] + D[I, J]
    loss = -(torch.log(nom + float_min) - torch.log(den + float_min)).sum()
    return loss

if __name__ == "__main__":
    eps = 1e-4
    n, d = 20, 2
    mu = 1e-2

    T = list(range(n))
    triplets = np.asarray([
        [h, o1, o2] if abs(h - o1) < abs(h - o2) else [h, o2, o1]
        for h, o1, o2 in itertools.product(T, T, T)
        if h != o1 and h != o2 and o1 != o2
    ])
    for i, (h, w, l) in enumerate(triplets):
        if np.random.uniform() < 0.20 and abs(abs(h - w) - abs(h - l)) <= 2:
            triplets[i] = [h, l, w]
    init = np.random.uniform(low=-eps, high=eps, size=(n, d))

    r = minimize(
        _ckl_x_loss,
        init,
        args=(init.shape, triplets, mu),
        method='L-BFGS-B',
        jac=True,
    )
    print(r.x.reshape(n, d))
```

I would expect graphs in the documentation and supplementary material to use a real triplet loss (preferably with real dataset(s)).

----

I've got some other style nits:

- Why do you use "accuracy improvement"? Does "0.1" on the x-axis mean a 10% improvement of a 0.1% improvement? What does each dot represent -- different datasets of different random seeds? Why show "accuracy improvement" instead of "accuracy"?
- What does "runtime improvement" mean? If FORTE has a value of 80%? I'm not even sure what to call 100%. I'd appreciate the raw times (maybe on a log scale).
  - If that's not possible because of datasets/etc, I think reformulating the graph to "slowdown factor" would make more sense to me (e.g., CKL took 1.2x more time than FORTE if FORTE went fastest).
- I'm not sure there's much overhead in "transform[ing the data] to pytorch tensors" as mentioned on the [user guide][ug] – conversion between the two is nearly instant because NumPy tensors and PyTorch Tensors share the same memory structure.

[ug]:https://cblearn.readthedocs.io/en/joss/user_guide/index.html#pytorch-backend-cpu-gpu
