### 2024-03-09 7245229
Added more explanation for linear subspace. Fixed the docs
