+++ docs/sg_execution_times.rst
@@ -1,46 +1,0 @@
-
-:orphan:
-
-.. _sphx_glr_sg_execution_times:
-
-
-Computation times
-=================
-**00:26.241** total execution time for 4 files **from all galleries**:
-
-.. container::
-
-  .. raw:: html
-
-    <style scoped>
-    <link href="https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/5.3.0/css/bootstrap.min.css" rel="stylesheet" />
-    <link href="https://cdn.datatables.net/1.13.6/css/dataTables.bootstrap5.min.css" rel="stylesheet" />
-    </style>
-    <script src="https://code.jquery.com/jquery-3.7.0.js"></script>
-    <script src="https://cdn.datatables.net/1.13.6/js/jquery.dataTables.min.js"></script>
-    <script src="https://cdn.datatables.net/1.13.6/js/dataTables.bootstrap5.min.js"></script>
-    <script type="text/javascript" class="init">
-    $(document).ready( function () {
-        $('table.sg-datatable').DataTable({order: [[1, 'desc']]});
-    } );
-    </script>
-
-  .. list-table::
-   :header-rows: 1
-   :class: table table-striped sg-datatable
-
-   * - Example
-     - Time
-     - Mem (MB)
-   * - :ref:`sphx_glr_generated_examples_plot_psychophysical_scales.py` (``../examples/plot_psychophysical_scales.py``)
-     - 00:14.731
-     - 0.0
-   * - :ref:`sphx_glr_generated_examples_triplet_formats.py` (``../examples/triplet_formats.py``)
-     - 00:05.455
-     - 0.0
-   * - :ref:`sphx_glr_generated_examples_ordinal_embedding.py` (``../examples/ordinal_embedding.py``)
-     - 00:04.205
-     - 0.0
-   * - :ref:`sphx_glr_generated_examples_small_embedding_benchmark.py` (``../examples/small_embedding_benchmark.py``)
-     - 00:01.850
-     - 0.0
