### 2024-03-12 9ad3fea
Synthetic Data Sampling from Manifold (#76)
* Added synthetic data sampling with LinearSubspace as the first one
* Added more explanation for linear subspace. Fixed the docs
