+++ cblearn/datasets/tests/test_linear_subspace.py
@@ -1,0 +1,224 @@
+import numpy as np
+import pytest
+from cblearn.datasets import LinearSubspace
+from functools import partial
+from scipy.spatial.distance import pdist, squareform
+
+# Subspace Dimension and Space Dimension tests
+
+
+def test_subspace_dimension_1():
+    manifold = LinearSubspace(subspace_dimension=1, space_dimension=3)
+    assert manifold.subspace_dimension == 1
+    points, _ = manifold.sample_points(num_points=10)
+    assert points.shape == (10, 3)
+    assert manifold.basis.shape == (2, 3)
+
+
+def subspace_dimension_greater_than_space_dimension(subspace_dimension, space_dimension, random_state):
+    _ = LinearSubspace(subspace_dimension=subspace_dimension, space_dimension=space_dimension, random_state=random_state)
+    # Add your actual test logic here
+
+
+@pytest.mark.parametrize("subspace_dimension, space_dimension, random_state", [
+    (1, 1, 1),
+    (1, 0, 1),
+    (2, 1, 1),
+    (4, 3, 1),
+])
+def test_subspace_dimension_greater_than_space_dimension(subspace_dimension, space_dimension, random_state):
+    with pytest.raises(ValueError):
+        subspace_dimension_greater_than_space_dimension(subspace_dimension, space_dimension, random_state)
+
+
+def invalid_subspace_dimension(subspace_dimension):
+    _ = LinearSubspace(subspace_dimension=subspace_dimension, space_dimension=3)
+
+
+@pytest.mark.parametrize("subspace_dimension", [-1, 0, "2", 1.5]
+                         )
+def test_invalid_subspace_dimension(subspace_dimension):
+    with pytest.raises(ValueError):
+        invalid_subspace_dimension(subspace_dimension)
+
+
+def invalid_space_dimension(space_dimension):
+    _ = LinearSubspace(subspace_dimension=1, space_dimension=space_dimension)
+
+
+@pytest.mark.parametrize("space_dimension", [-1, 0, "2", 1.5]
+                         )
+def test_invalid_space_dimension(space_dimension):
+    with pytest.raises(ValueError):
+        invalid_space_dimension(space_dimension)
+
+# # Noise and Sampling Function Tests
+
+
+@pytest.mark.parametrize("subspace_dimension, space_dimension", [
+    (1, 2),
+    (2, 2),
+    (4, 4),
+    (1, 3),
+    (3, 10)
+])
+def test_no_random_state(subspace_dimension, space_dimension):
+    manifold = LinearSubspace(subspace_dimension=subspace_dimension, space_dimension=space_dimension, random_state=None)
+    points, _ = manifold.sample_points(num_points=10)
+    assert points.shape == (10, space_dimension)
+
+
+@pytest.mark.parametrize("subspace_dimension, space_dimension, random_state", [
+    (1, 2, 1),
+    (2, 2, 2),
+    (4, 4, 10),
+    (1, 3, 1),
+    (3, 10, 1)
+])
+def test_no_noise(subspace_dimension, space_dimension, random_state):
+    manifold1 = LinearSubspace(
+        subspace_dimension=subspace_dimension,
+        space_dimension=space_dimension,
+        random_state=random_state)
+    manifold1._create_manifold()
+    manifold2 = LinearSubspace(
+        subspace_dimension=subspace_dimension,
+        space_dimension=space_dimension,
+        random_state=random_state)
+    manifold2._create_manifold()
+    # Check that basis is the same
+    np.testing.assert_array_equal(manifold1.basis, manifold2.basis)
+
+
+@pytest.mark.parametrize("subspace_dimension, space_dimension, noise", [
+    (1, 2, 'normal'),
+    (2, 2, 'laplace'),
[…]
