### 2024-03-13 91e7846
Update torch speed test figure and description in docs
