### 2024-03-13 6cb9654
Update torch speedtest in supplementary to minimal CKL
