+++ paper/references.bib
@@ -206,6 +206,12 @@
 	pages = {19--19},
 	doi = {10.1167/12.3.19},
 }
+@article{kingma2014adam,
+  title={Adam: A method for stochastic optimization},
+  author={Kingma, Diederik P and Ba, Jimmy},
+  journal={arXiv preprint arXiv:1412.6980},
+  year={2014}
+}
 
 @article{hebartRevealingMultidimensionalMental2020,
 	title = {Revealing the multidimensional mental representations of natural objects underlying human similarity judgements},
