+++ paper/supplementary.md
@@ -44,7 +44,7 @@ An additional advantage of GPU implementations is that they require no explicit 
     \label{fig:time-per-triplets_gpu}](images/time-per-triplets_gpu.pdf){width=50%}
 
 There are various explanations for the speed disadvantage of our pytorch implementations. On the one hand, it may be due to the overhead of converting between numpy and pytorch and calculating the gradient (AutoGrad). On the other hand, it can also be due to the optimizer or the selected hyperparameters. 
-To get a first impression of these factors, we have built minimal examples of the CKL algorithm [@tamuz_adaptively_2011] and estimated 2D embeddings of the Vogue Cover dataset [@heikinheimo2013crowd]. \autoref{fig:torch-speedtest} shows the runtimes and triplet accuracies on a standard laptop. The small markers show runs with different initialization and the bold markers the respective means. The CKL implementation of `cblearn` is approximately three times slower than the minimal version, probably due to data validation and conversion overheads. If the gradient is not provided directly but calculated automatically with PyTorch's AutoGrad functions, the minimal example runs ~11 times slower. The most severe impact has changing the optimization algorithm to stochastic optimization (*Adam*, lr=10).  However, it can be assumed in accordance with the results in previous sections, that this overhead is compensated with increasing dataset size. 
+To get a first impression of these factors, we have built minimal examples of the CKL algorithm [@tamuz_adaptively_2011] and estimated 2D embeddings of the Vogue Cover dataset [@heikinheimo2013crowd]. \autoref{fig:torch-speedtest} shows the runtimes and triplet accuracies on a standard laptop. The small markers show runs with different initialization and the bold markers the respective means. The CKL implementation of `cblearn` is approximately three times slower than the minimal version, probably due to data validation and conversion overheads. If the gradient is not provided directly but calculated automatically with PyTorch's AutoGrad functions, the minimal example runs ~9 times slower. The most severe impact has changing the optimization algorithm to stochastic optimization (*Adam*, lr=10).  However, it can be assumed in accordance with the results in previous sections, that this overhead is compensated with increasing dataset size. 
 
 ![The runtime and error for different optimization methods in minimal CKL implementations. `cblearn`'s CKL implementation is shown for reference.\label{fig:torch-speedtest}](
     images/torch_speedtest_triplets.pdf
