### 2024-04-15 #6139 @stsievert
@dekuenstle your update looks pretty good to me. I've got one substantive complaint on Fig. 3 in the supplementary, and some suggestions on the figures. I agree with you, the plots have become more insightful and more interesting.

To me, I think Fig. 3 in the supplementary is the most interesting figure. But, I had a hard time telling exactly what's different between the different implementations. I've looked a little closer, and have noticed a difference with some hyperparameters between the "cblearn" implementation and the two minimal SciPy implementation ("minimal SciPy w/ L-BFGS-B" & "minimal SciPy w/ AutoGrad"). It looks like  `mu=1` for the two minimal SciPy implementations<sup>*</sup> and `mu=0` [for cblearn implementation][1]. That may influence the speed results, which makes me question the statement "cblearn is approximately three times slower than the minimal version, probably due to data validation and conversion overhead". The speed differences (3x, 9x and 547x) seem large; do those speed differences hold up to slightly larger datasets too?

<sup>* See `mu = 1` at the beginning of the function `optimize` in [pytorch_speedtest_triplets.ipynb][pst]</sup>

[1]:https://github.com/cblearn/cblearn/blob/391dd152bec5069e12813a178cf20d25fb5a5ed1/cblearn/embedding/_ckl.py#L61
[pst]:https://github.com/cblearn/cblearn-benchmark/blob/b40a517ca7090d7a1d813e5e7f61cf1f4af9cd42/scripts/pytorch_speedtest_triplets.ipynb

- Documentation: I think it'd be a good idea to put a lot of the figures/Q&A in the paper/supplementary into the documentation. The cblearn users might find those interesting (e.g., supplementary code example to "getting started" guide).
- Paper: You've got a typo with `*cblearn*`, I see `... all implementations, but *cblearn* is many times faster ...` in the PDF
- Supplementary Fig. 1 in the supplementary could be improved, I find the x-axis meaningless. I think it'd better if there were a histogram for each triplet class (e.g., CKL) over error/slowdown (x=error/slowdown, y=frequency, color=triplet class).
- Supplementary Fig. 3: I'd appreciate if test error/times were separated (or split into another graph like Figure 1 and plotted per the above recommendation).
  - Supplementary: it's probably a good idea to plot the median time (not the mean time) to filter out runs when the OS/etc gets in the way. Those (few?) slow values will skew the estimation of "3x" and "11x" because the mean is sensitive to outliers, the median isn't.
- Supplementary: You've got some code left over from your Adam LR toy example:
  - The figure `images/adam_lr_triplets.pd` isn't showing, I see a `width=50%` in the PDF.
  - "This sensitivity [to LR] is demonstrated in Figure , where the learning rate"
- Another wording complaint: with "... additional advantage of GPU implementations is that they
require no explicit gradient definition". I'd make it clear that this is a PyTorch feature, not a GPU feature.
