### 2024-05-08 b741681
Added a line sampling method
