### 2024-05-13 792a12e
Fix typos / wording in supplementary
