+++ paper/references.bib
@@ -212,6 +212,14 @@
   journal={arXiv preprint arXiv:1412.6980},
   year={2014}
 }
+@inproceedings{heikinheimo2013crowd,
+  title={The crowd-median algorithm},
+  author={Heikinheimo, Hannes and Ukkonen, Antti},
+  booktitle={Proceedings of the AAAI Conference on Human Computation and Crowdsourcing},
+  volume={1},
+  pages={69--77},
+  year={2013}
+}
 
 @article{hebartRevealingMultidimensionalMental2020,
 	title = {Revealing the multidimensional mental representations of natural objects underlying human similarity judgements},
