### 2024-05-14 0ca819d
CI install loe from archive
