+++ .github/workflows/test.yml
@@ -35,7 +35,7 @@ jobs:
         #    see .flake8 config file for selected/ignored rules.
         #    warnings can be found in the action logs
 
-  docs: 
+  docs:
     runs-on: ubuntu-latest
     steps:
     - uses: actions/checkout@v3
@@ -55,7 +55,7 @@ jobs:
   test:
     strategy:
       matrix:
-        python-version: 
+        python-version:
           - "3.9"
           - "3.10"
           - "3.11"
@@ -70,6 +70,11 @@ jobs:
         cache-dependency-path: setup.cfg
     - name: Setup R
       uses: r-lib/actions/setup-r@v2
+    - name: Install R dependencies
+      run: |
+        # 2024-05-14: loe is not available from CRAN, we have to fallback to the archive.
+        wget https://cran.r-project.org/src/contrib/Archive/loe/loe_1.1.tar.gz
+        R CMD INSTALL ./loe_1.1.tar.gz
     - name: Install package with dependencies
       run: |
         python3 -m pip install --upgrade pip
@@ -85,4 +90,3 @@ jobs:
         file: ./coverage.xml
         flags: unittests
         env_vars: OS,PYTHON
-      
