+++ .github/workflows/build-paper-pdf.yaml
@@ -28,7 +28,7 @@ jobs:
             --resource-path=paper/
             paper/supplementary.md
       - name: Upload
-        uses: actions/upload-artifact@v3
+        uses: actions/upload-artifact@v4
         with:
           name: paper-pdf
           # This is the output path where Pandoc will write the compiled
