### 2024-05-14 6d5a6a7
Fix typo and style in supplementary
