+++ paper/supplementary.md
@@ -1,7 +1,7 @@
 ---
 title: |
        | Supplementary Material
-       | *cblearn: Comparison-based Machine Learning in Python*
+       | `cblearn`: Comparison-based Machine Learning in Python*
 author:
   - |
     | David-Elias Künstle and Ulrike von Luxburg
@@ -14,8 +14,8 @@ date: 22 September 2023
 We generated embeddings of comparison-based datasets to measure runtime and triplet error as a small empirical evaluation of our ordinal embedding implementations.
 We compared various CPU and GPU implementations in `cblearn` with third-party implementations in *R* [`loe` @terada_local_2014], and *MATLAB* [@van_der_maaten_stochastic_2012].
 In contrast to synthetic benchmarks [e.g., @vankadara_insights_2020], we used the real-world datasets
-that can be accessed through *cblearn*, converted to triplets. The embeddings were arbitrarily chosen to be 2D.
-Every algorithm runs once per dataset on a compute node (8 cores of a Intel\textregistered Xeon \textregistered Gold 6240; 96GB RAM; NVIDIA RTX 2080ti) with a run-time limit of 24 hours. Some runs did fail by exceeding those constraints: our *FORTE* implementation failed by an out of memory error on the *imagenet-v2* dataset. The *MATLAB* implementation of *tSTE* timed out on *things* and *imagenet-v2* datasets. The *R* implementation of *SOE* on the *imagenet-v2* dataset by an "unsupported long vector" error, caused by the large size of the requested embedding.
+that can be accessed through `cblearn`, converted to triplets. The embeddings were arbitrarily chosen to be 2D.
+Every algorithm runs once per dataset on a compute node (8 cores CPU; 96GB RAM; NVIDIA RTX 2080ti) with a run-time limit of 24 hours. Some runs did fail by exceeding those constraints: our FORTE implementation failed by an out of memory error on the `imagenet-v2` dataset. The *MATLAB* implementation of tSTE timed out on `things` and `imagenet-v2` datasets. The *R* implementation of SOE on the `imagenet-v2` dataset by an "unsupported long vector" error, caused by the large size of the requested embedding.
 
 The benchmarking scripts and results are publicly available in a separate repository[^1].
 
@@ -24,21 +24,20 @@ The benchmarking scripts and results are publicly available in a separate reposi
 ## Is there a "best" estimator?
 
 
-Comparing all ordinal embedding estimators in `cblearn`, *SOE*, *CKL*, *GNMDS*, and *tSTE* were performing about equally well in both runtime and accuracy (\autoref{fig:performance-per-algorithm_cblearn}).
-The GPU implementations are slower on the tested datasets and for *SOE* and *GNMDS* noticeably less accurate.
+Comparing the ordinal embedding estimators in `cblearn`, SOE, CKL, GNMDS, and tSTE were performing about equally well in both runtime and accuracy (\autoref{fig:performance-per-algorithm_cblearn}).
+The GPU implementations are slower on the tested datasets and for SOE and GNMDS noticeably less accurate.
 
 ![\label{fig:deltaerror-per-algorithm_cblearn-all}](./images/deltaerror-per-algorithm_cblearn-all.pdf){width=45%}
 ![\label{fig:deltatime-per-algorithm_cblearn-all}](images/deltatime-per-algorithm_cblearn-all.pdf){width=45%}
 \begin{figure}
-\caption{The triplet error and runtime per estimator and dataset, relative to the mean error or the fastest run. Thin lines show runs on the different datasets; the thick lines indicate the respective median. With the exception of *STE*, all CPU algorithms are able to embed the triplets similarly well. There are just minor differences in the runtime of the CPU implementations. The GPU implementations are usually significantly slower on the data sets used.  
[…]
