+++ paper/references.bib
@@ -220,6 +220,14 @@
   pages={69--77},
   year={2013}
 }
+@article{paszke2019pytorch,
+  title={Pytorch: An imperative style, high-performance deep learning library},
+  author={Paszke, Adam and Gross, Sam and Massa, Francisco and Lerer, Adam and Bradbury, James and Chanan, Gregory and Killeen, Trevor and Lin, Zeming and Gimelshein, Natalia and Antiga, Luca and others},
+  journal={Advances in neural information processing systems},
+  volume={32},
+  year={2019}
+}
+
 
 @article{hebartRevealingMultidimensionalMental2020,
 	title = {Revealing the multidimensional mental representations of natural objects underlying human similarity judgements},
