### 2024-05-14 e82b354
Reference scipy, numpy, pytorch
