### 2024-05-14 #6139 @dekuenstle
Dear @stsievert ,

thank you once again for your detailed comments. Let me respond point-by-point:

> mu = 0 vs mu = 1

Thank you for this excellent catch. I did rerun the experiment with `mu=0` in all conditions and updated the plots. Now, cblearn is just 20% slower than the minimal scipy version. The other conditions remain substantially slower.

![](https://github.com/cblearn/cblearn-benchmark/blob/main/plots/torch_speedtest_triplets.png?raw=true)

>  do those speed differences hold up to slightly larger datasets too?

I have repeated the experiment on the slightly larger `nature` dataset (3x times  the triplets, 2x the objects). I got approximately the same runtime differences between scipy and torch conditions (median of 10 runs):
```
                     time     error
method
cblearn          0.514675  0.163562
scipy            0.586388  0.162444
scipy+autograd   3.295194  0.162742
torch           99.711365  0.164754
```

However, Supplementary Figure 2 indicates that this difference will eventually decrease with increasing dataset size and I highlight this observation in the text.

> extend documentation with figures, code example, q&a from the paper

I completely agree that extending the documentation must be a longterm goal.

With the help of your feedback, I already improved the documentation substantially. For example, I added a quick start guide including code example and some figures from the supplementary to the user guide. You find this version of the documentation [here](https://cblearn.readthedocs.io/en/joss/getting_started/index.html#quick-start). It will be relased once I merge back the JOSS branch to main.

However, I prefer not to add all benchmark plots to the documentation. They mainly describe the status quo at the time of this paper. Updating the benchmark (and the plots) might be worth for substantial changes of `cblearn` but will not be feasible on a regular basis.

> paper typo

Thank you, I fixed it.

> supplementary figure 1: meaningless x-axis -> histogram

I agree that there are many variations to visualize this data. I chose a lineplot over a histogram because I want to highlight the median performance, not distribution.

> supplementary figure 3: separate test time/error

I appreciate your recommendation to split the plot. In this case, prefer the current version which allows to relate both dimensions and provides both the median and distribution information.

> supplementary figure 3: median instead of mean

Thank you, I followed your suggestion and updated the plot.

> Figure images/adam_lr_triplets.pdf not showing / reference broken

Thank you, I fixed this error and updated the [supplemenary pdf](https://github.com/cblearn/cblearn/blob/joss/paper/supplementary.pdf).

> Wording: "Another wording complaint: with "... additional advantage of GPU implementations is that they
require no explicit gradient definition". I'd make it clear that this is a PyTorch feature, not a GPU feature."

I completely agree, that this is a pytorch feature. I updated the respective sentence.
In addition, I cited the pytorch project both in the main paper and the supplementary.
