+++ paper/references.bib
@@ -9,14 +9,21 @@
   pages = {1--6}
 }
 
-@article{sauer2023seeing,
-  title={Seeing the future of progressive glasses: a new perceptual approach to spectacle lens design},
-  author={Sauer, Yannick and K{\"u}nstle, David-Elias and Wichmann, Felix and Wahl, Siegfried},
-  year={2023},
-  publisher={PsyArXiv},
-  doi={10.31234/osf.io/pge5n}
+@article{fsauerObjectiveMeasurementApproach2024,
+  title = {An Objective Measurement Approach to Quantify the Perceived Distortions of Spectacle Lenses},
+  author = {Sauer, Yannick and K{\"u}nstle, David-Elias and Wichmann, Felix A. and Wahl, Siegfried},
+  year = {2024},
+  month = feb,
+  journal = {Scientific Reports},
+  volume = {14},
+  number = {1},
+  pages = {3967},
+  issn = {2045-2322},
+  doi = {10.1038/s41598-024-54368-3},
+  urldate = {2024-02-19},
+  copyright = {All rights reserved},
+  langid = {english},
 }
-
 @article{huber2024tracing,
   title={Tracing Truth Through Conceptual Scaling: Mapping People’s Understanding of Abstract Concepts},
   author={Huber, Lukas S and K{\"u}nstle, David-Elias and Reuter, Kevin},
@@ -206,11 +213,20 @@
 	pages = {19--19},
 	doi = {10.1167/12.3.19},
 }
-@article{kingma2014adam,
-  title={Adam: A method for stochastic optimization},
-  author={Kingma, Diederik P and Ba, Jimmy},
-  journal={arXiv preprint arXiv:1412.6980},
-  year={2014}
+
+@inproceedings{kingma2014adam,
+  author       = {Diederik P. Kingma and
+                  Jimmy Ba},
+  editor       = {Yoshua Bengio and
+                  Yann LeCun},
+  title        = {Adam: {A} Method for Stochastic Optimization},
+  booktitle    = {3rd International Conference on Learning Representations, {ICLR} 2015,
+                  San Diego, CA, USA, May 7-9, 2015, Conference Track Proceedings},
+  year         = {2015},
+  url          = {http://arxiv.org/abs/1412.6980},
+  timestamp    = {Thu, 25 Jul 2019 14:25:37 +0200},
+  biburl       = {https://dblp.org/rec/journals/corr/KingmaB14.bib},
+  bibsource    = {dblp computer science bibliography, https://dblp.org}
 }
 @inproceedings{heikinheimo2013crowd,
   title={The crowd-median algorithm},
