### 2024-05-30 #6139 @mbarzegary
@dekuenstle given the green light of the reviewers, we will now work towards processing this for acceptance in JOSS. So please
* Merge my PR with some minor edits
* Work on the author's points of the final checklist I created ☝️
* Add `sauer2023seeing` reference as it is missing now and prevents the paper from rendering correctly
* Fix the missing DOIs issue reported by the editorial bot

Important note: in the PR, I fixed the link to the supplementary material (which exists in the `joss` branch rather than `main`). Please make sure that the branch stays alive in the future otherwise the link becomes inaccessible.
