### 2024-05-31 1cdf29d
Update reference in paper.md
sauer2023 -> sauer2024
(preprint -> published paper)
