+++ paper/references.bib
@@ -138,12 +138,10 @@
 	doi = {10.1167/3.8.5}
 }
 
-@inproceedings{agarwal_generalized_2007,
-	title = {Generalized non-metric multidimensional scaling},
-	booktitle = {Artificial {Intelligence} and {Statistics} (AISTATS)},
-	author = {Agarwal, Sameer and Wills, Josh and Cayton, Lawrence and Lanckriet, Gert and Kriegman, David and Belongie, Serge},
-	year = {2007},
-}
+@InProceedings{agarwal_generalized_2007,
+		title =  {Generalized Non-metric Multidimensional Scaling},  
+		author =  {Agarwal, Sameer and Wills, Josh and Cayton, Lawrence and Lanckriet, Gert and Kriegman, David and Belongie, Serge},  booktitle =  {Proceedings of the Eleventh International Conference on Artificial Intelligence and Statistics},  pages =  {11--18},  year =  {2007},  editor =  {Meila, Marina and Shen, Xiaotong},  volume =  {2},  series =  {Proceedings of Machine Learning Research},  address =  {San Juan, Puerto Rico},  month =  {21--24 Mar},  publisher =    {PMLR},  pdf =  {http://proceedings.mlr.press/v2/agarwal07a/agarwal07a.pdf},  url =  {https://proceedings.mlr.press/v2/agarwal07a.html}}
+
 
 @inproceedings{terada_local_2014,
 	title = {Local ordinal embedding},
