### 2024-05-31 #6139 @dekuenstle
@stsievert: Once again, many thanks for your detailed comments! I think they substantially improved the package documentation and paper!

@mbarzegary
I merged your pull requests and fixed the `sauer2023seeing` reference by updating it from the preprint to the published version.
 As far as I know, I cannot add the outstanding DOIs because the papers were published at machine-learning conferences or journals that do not assign DOIs. But I will check again.

This is the list of missing DOIs, ordered by publisher
```
JMLR (PMLR)
 metric-learn: Metric Learning Algorithms in Python
 Scikit-learn: Machine Learning in Python

NeurIPS Proceedings
 NEXT: A System for Real-World Development, Evaluat...
 Finite Sample Prediction and Recovery Bounds for O...
 Pytorch: An imperative style, high-performance dee...
 Foundations of Comparison-Based Hierarchical Clust...
 Near-optimal comparison based clustering
 Landmark Ordinal Embedding

AISTATS
 Generalized non-metric multidimensional scaling

ICML (PMLR)
 Local ordinal embedding
 Adaptively learning the crowd kernel
 Comparison-Based Random Forests
 Multiview Triplet Embedding: Learning Attributes i...
 Scaling up ordinal embedding: A landmark approach

ICLR
 Adam: A Method for Stochastic Optimization

AAAI
 The crowd-median algorithm

Conference on learning theory (PMLR)
 Learning combinatorial functions from pairwise com...
```

Next, I will work on the author tasks above.
