### 2024-05-31 #6139 @mbarzegary
@dekuenstle I see. So, I think we can skip the DOIs if they are not available at all.
Also, I noticed that the rendering of the "Vankadara et al. (2021)" citation is incorrect in the text (twice) and not in line with the rest. can you please fix these ones too?
