### 2024-05-31 #6139 @dekuenstle
@mbarzegary I will have another proof read over the references to check for consistency.

However, I think I do not see the rendering issue that you point out. Could you help me out, please?
```
Inspired by the work of Vankadara et al. (2021), ...
...focusing on large data sets, can be found in Vankadara et al. (2021).
```
