### 2024-05-31 #6139 @mbarzegary
@dekuenstle the consistency problem for the mentioned reference is due to lack of square brackets. Please replace both occurrences of `@vankadara_insights_2020` to `[@vankadara_insights_2020]` to fix the issue.
