### 2024-05-31 #6139 @dekuenstle
@mbarzegary I'm not sure why there should be brackets. I intentionally use in-text citations (i.e., omit brackets) while "referring to the author as part of a sentence" [(JOSS Author Guides)](https://joss.readthedocs.io/en/latest/paper.html#citations).
Is there a convention to avoid in-text citations in JOSS? Of course, I can rewrite those sentences and add brackets.
