### 2024-06-01 #6139 @mbarzegary
@dekuenstle my bad, sorry. Yes, it's correct to use this style in a sentence. Please proceed with completing the author's points in the checklist.
