+++ CHANGELOG.md
@@ -1,8 +1,17 @@
 # Changelog
 
 ## Upcoming
+
+## 0.3
+
+- Feature: JOSS paper
+- Feature: Quickstart guide in documentation
 - Feature: Data point sampling from manifolds.
-# 0.2.0
+- Improvement: Extended documentation 
+- Improvement: Filter invalid responses in datasets
+- Improvement: Full compatibility to sklearn estimator tests
+  
+## 0.2
 
 - Improvement: Extended documentation
 - Feature: `embedding.estimate_dimensionality_cv` function (Künstle et al., 2022)
@@ -11,17 +20,18 @@
 - Fix: Minor errors in the unit tests
 - Others: Updated dependencies
 
-## 0.1.2
+## 0.1
+### 0.1.2
 
 - support python 3.11
 - update core dependencies
   
-## 0.1.1
+### 0.1.1
 
 - Minor fixes in the documentation.
 - Adapt loading of food and imagenet dataset to solve problems caused by changes in externally hosted files
   
-## 0.1.0
+### 0.1.0
 
 - Support python 3.9 and 3.10.
 - Introduce semantic versioning
