### 2024-06-03 fa34139
Remove unneeded folder in docs
