### 2024-06-03 065e87e
Restyle docs with logo and pydata theme
