### 2024-06-03 a460e5a
Shrink gh-action workflow names for badges
