+++ .github/workflows/release.yml
@@ -1,6 +1,6 @@
 # Based on:
 # https://packaging.python.org/en/latest/guides/publishing-package-distribution-releases-using-github-actions-ci-cd-workflows/
-name: Publish Python 🐍 distributions 📦 to PyPI and TestPyPI
+name: Release
 
 on:
   push:
