### 2024-06-03 30fc42c
Add logo to readme
