+++ README.md
@@ -1,19 +1,22 @@
-# cblearn
+<h1 align="center">
+<img src="https://raw.githubusercontent.com/cblearn/cblearn/main/docs/logo-light.svg" width="300">
+</h1><br>
+
 ## Comparison-based Machine Learning in Python
 [![PyPI version](https://img.shields.io/pypi/v/cblearn.svg)](https://pypi.python.org/pypi/cblearn)
 [![Documentation](https://readthedocs.org/projects/cblearn/badge/?version=stable)](https://cblearn.readthedocs.io/en/stable/?badge=stable)
 [![Test status](https://github.com/cblearn/cblearn/actions/workflows/test.yml/badge.svg?branch=main)](https://github.com/cblearn/cblearn/actions/workflows/test.yml)
 [![Test Coverage](https://codecov.io/gh/cblearn/cblearn/branch/master/graph/badge.svg?token=P9JRT6OK6O)](https://codecov.io/gh/cblearn/cblearn)
 
-Comparison-based Learning algorithms are the Machine Learning algorithms to use when training data contains similarity comparisons ("A and B are more similar than C and D") instead of data points. 
+Comparison-based learning algorithms are the machine learning algorithms to use when the training data contains similarity comparisons ("A and B are more similar than C and D") instead of data points. 
 
 Triplet comparisons from human observers help model the perceived similarity of objects.
-These human triplets are collected in studies, asking questions like 
-"Which of the following bands is most similar to Queen?" or 
-"Which color appears most similar to the reference?".
+These human triplets are collected in studies that ask questions such as 
+"Which of the following bands looks most like Queen?" or 
+"Which color is most similar to the reference?
 
 This library provides an easy-to-use interface for comparison-based learning algorithms.
-It plays hand-in-hand with scikit-learn:
+It works hand in hand with scikit-learn:
 
 ```python
 from sklearn.datasets import load_iris
@@ -44,37 +47,6 @@ Install cblearn as described [here](https://cblearn.readthedocs.io/en/stable/ins
 
 Find a theoretical introduction to comparison-based learning, the datatypes, 
 algorithms, and datasets in the [User Guide](https://cblearn.readthedocs.io/en/stable/user_guide/index.html).
-
-## Features
-
-### Datasets
-
-*cblearn* provides utility methods to simplify the loading and conversion
-of your comparison datasets. In addition, some functions download and load multiple real-world comparisons.
-
-| Dataset  | Query | #Object | #Response | #Triplet |
-| --- | --- | ---:| ---:| ---:|
-| Vogue Cover | Odd-out Triplet | 60 | 1,107 | 2,214 | 
-| Nature Scene | Odd-out Triplet | 120 | 3,355 | 6,710 | 
-| Car | Most-Central Triplet | 60 | 7,097 | 14,194 | 
-| Material | Standard Triplet | 100 | 104,692 |104,692 | 
-| Food | Standard Triplet | 100 | 190,376 |190,376 | 
-| Musician | Standard Triplet | 413 | 224,792 |224,792 | 
-| Things Image Testset | Odd-out Triplet | 1,854 | 146,012 | 292,024 | 
-| ImageNet Images v0.1 | Rank 2 from 8 | 1,000 | 25,273 | 328,549 | 
-| ImageNet Images v0.2 | Rank 2 from 8 | 50,000 | 384,277 | 5M | 
-
-
-### Embedding Algorithms
-
-| Algorithm                   | Default | Pytorch (GPU) | Reference Wrapper |
-| --------------------------- |  :---:  | :-----------: | :---------------: |
-| Crowd Kernel Learning (CKL) | X       | X             |                   |
-| FORTE                       |         | X             |                   |
[…]
