### 2024-06-03 896a8d8
Shrink readme content
