### 2024-06-03 4202eaa
Shrink readme further:
