### 2024-06-06 1bf3a5c
Fix typos, citation, acknowledgment
