+++ paper/paper.md
@@ -35,8 +35,7 @@ The `cblearn` package implements comparison-based machine learning algorithms
 and routines to process comparison-based data in Python.
 Comparison-based learning algorithms are used when only comparisons of similarity between data points are available, but no explicit similarity scores or features.
 For example, humans struggle to assign *numeric* similarities to apples, pears, and bananas.
-Still, they can easily *compare* the similarity of pears and apples with the similarity of apples and bananas---
-pears and apples usually appear more similar.
+Still, they can easily *compare* the similarity of pears and apples with the similarity of apples and bananas---pears and apples usually appear more similar.
 There exist comparison-based algorithms for most machine learning tasks,
 like clustering, regression, or classification [e.g., @balcan2016learning; @heikinheimo2013crowd; @perrot_near-optimal_2020];
 The most frequently applied algorithms, however, are the so-called ordinal embedding algorithms
@@ -52,8 +51,7 @@ various stimuli objectively
 # Statement of need
 
 This work presents `cblearn`, an open-source Python package for comparison-based learning. Unlike related packages,
-`cblearn` goes beyond specific algorithm implementations to provide an ecosystem for
-for comparison-based data with access to several real-world datasets
+`cblearn` goes beyond specific algorithm implementations to provide an ecosystem for comparison-based data with access to several real-world datasets
 and a collection of algorithm implementations.
 `cblearn` is fast and user-friendly for applications but flexible for research on new algorithms and methods.
 The package integrates well into the scientific Python ecosystem; for example, third-party functions for cross-validation or hyperparameter tuning of `scikit-learn` estimators can typically be used with `cblearn` estimators.
@@ -67,7 +65,7 @@ for processing and converting the comparison data in all its varieties (`cblearn
 The atomic datum in comparison-based learning is the quadruplet,
 a comparison of the similarity $\delta$ between two pairs $(i, j)$ and $(k, l)$,
 for example, asserting that $\delta(i, j) < \delta(k, l)$.
-Alternative comparisons like the triplet, where $i == l$, can be reduced to one or more quadruplets.
+Another popular comparison query, the triplet, can be reduced to a quadruplet with $i == l$.
 Comparison-based learning algorithms estimate classes, clusters, or metrics to fulfill as many quadruplets as possible.
 In ordinal embedding, for example, the problem is to find  $x_i, x_j, x_k, x_l \in \mathbb{R}^d$
 s.t. $\left\lVert x_i - x_j \right\rVert_2 < \left\lVert x_k - x_l \right\rVert_2  \Leftrightarrow \delta(i, j) < \delta(k, l)$.
@@ -75,8 +73,8 @@ s.t. $\left\lVert x_i - x_j \right\rVert_2 < \left\lVert x_k - x_l \right\rVert_
 Besides triplets and quadruplets, there are many ways to ask for comparisons.
 Some tasks ask for the "odd-one-out", the "most-central" object, or the two most similar objects to a reference. `cblearn` can load these different queries and convert them to triplets, ready for subsequent embedding or clustering tasks.
 
-Different data types can store triplets, and  `cblearn` converts them internally.
-A 2D array with three columns for the object indices $(i, j, k)$ stores a triplet per row. In some applications, it is comfortable to separate the comparison "question" and "response", which leads to an additional list of labels $1$, if $\delta(i, j) \le \delta(i, k)$, and $-1$, if $\delta(i, j) > \delta(i, k)$.
+Different data types can store triplets and  `cblearn` converts them internally.
[…]
