### 2024-06-06 d5b1a12
Fix wording and typos in supplementary
