### 2024-06-06 8b77bbe
Restructure sentence to fix citation rendering
