### 2024-06-06 d5177e7
Add JOSS paper and improvements (#78)
* Create references.bib
* Create paper.md
* Create build-paper-pdf.yaml
* add supplementaries
* extend unit tests
* extend documentation

Co-authored-by: Mojtaba Barzegari <40744245+mbarzegary@users.noreply.github.com>
