+++ examples/ordinal_embedding.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 r"""
 .. _example_ordinal_embedding:
+
 Ordinal Embedding
 =================
 

[…]
