### 2024-06-06 348ed0d
Update image link in readme
