### 2024-06-06 #6139 @dekuenstle
- [x]   Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.  [**0.3.0** ](https://pypi.org/project/cblearn/)
- [x]   Archive the release on Zenodo/figshare/etc and post the DOI here.  [**10.5281/zenodo.11410206**](https://zenodo.org/doi/10.5281/zenodo.11410206).
