### 2024-06-11 #6139 @mbarzegary
@dekuenstle you have duplicate entries for two references (`harris_array_2020` and `heikinheimo2013crowd`). please remove the redundant ones so that we can proceed with the acceptance procedure. also, for the latter, [10.1609/hcomp.v1i1.13079](https://doi.org/10.1609/hcomp.v1i1.13079) seems to be a valid DOI. please check this too.
