### 2024-06-12 #6139 @dekuenstle
@mbarzegary I removed the duplicates. One of the `heikinheimo2013crowd` entries already contained the DOI.
