### 2024-06-12 #6139 @crvernon
🥳 Congratulations on your new publication @dekuenstle! Many thanks to @mbarzegary for editing and @haniyeka, @sherbold, and @stsievert for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
